### 2024-01-19 #6248 @editorialbot
[PRE REVIEW]: MultichannelGPR: A MATLAB tool for GPR data processing
**Submitting author:** @tinawunderlich (<a href="http://orcid.org/0000-0003-2275-704X">Tina Wunderlich</a>)
**Repository:** https://github.com/tinawunderlich/MultichannelGPR
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.4
