### 2024-01-20 5e32323
improved Layerpicking with scrollbars for long profiles and different picking geometries
