+++ GUIs/plot_layerpicking.m
@@ -36,7 +36,6 @@ S.prof_old=1;
 
 % get screensize
 S.siz = get( 0, 'Screensize' );
-S.fh.Position=S.siz;  % default: screensize
 
 % Add the UI components
 S=addcomponents(S);
@@ -53,16 +52,24 @@ S.fh.Visible='on';
             'menubar','none',...
             'name','Layer picking',...
             'numbertitle','off',...
-            'resize','on','Visible','off','SizeChangedFcn',@resizeui);
+            'resize','on','Visible','off','SizeChangedFcn',@resizeui,'Renderer','painters');
         
+        % --
+        % upper plot for comparison
+        % scrollbar:
+        S.panel1=uipanel('Parent',S.fh);
+        S.panel1a=uipanel('Parent',S.panel1);
+        set(S.panel1,'Position',[210 80+S.siz(4)/2+80 S.siz(3)-220 S.siz(4)/2-100],'Units','pixels'); % panel for cutting frame with slider
+        set(S.panel1a,'Position',[210 80+S.siz(4)/2+80 2000 S.siz(4)/2-100],'Units','pixels'); % panel inside cutting frame with image
+
+        S.slider1=uicontrol('Style','Slider','Parent',S.fh,'Units','pixels','Position',[210 80+S.siz(4)/2+80 S.siz(3)-220 0.05],'Value',0.5,'Callback',@slider_callback1);
         
-        S.ax1 = axes('unit','pix',...
-            'position',[200 100 S.siz(3)-200 S.siz(4)-500]);
+        S.ax1 = axes('OuterPosition',[0.02 0.02 0.98 0.98],'Units','normalized','Parent',S.panel1a);
         % make initial plot
         if isnan(S.z)
-            S.Plot1 = imagesc(S.x{1},S.t,S.radargrams{1});
+            S.Plot1 = imagesc(S.ax1,S.x{1},S.t,S.radargrams{1});
         else
-            S.Plot1 = imagesc(S.x{1},S.z,S.radargrams{1});
+            S.Plot1 = imagesc(S.ax1,S.x{1},S.z,S.radargrams{1});
         end
         colormap(flipud(gray));
         xlabel('x [m]')
@@ -73,17 +80,29 @@ S.fh.Visible='on';
             ylabel('t [ns]')
             axis ij
         end
-       set(S.ax1,'DataAspectratio',[0.5 1 1])
-        title(['For comparison: Profile #'])
+        set(S.ax1,'DataAspectratio',[0.5 1 1])
         
+        S.title1 = uicontrol(S.fh,'style','text',...
+            'unit','pix',...
+            'position',[210 900 200 25],'String','For comparison: Profile #','FontWeight','bold');
         
-        S.ax2 = axes('unit','pix',...
-            'position',[200 500 S.siz(3)-200 S.siz(4)-500]);
+
+        %--
+        % Lower plot for picking:
+        % scrollbar:
+        S.panel2=uipanel('Parent',S.fh);
+        S.panel2a=uipanel('Parent',S.panel2);
+        set(S.panel2,'Position',[210 120 S.siz(3)-220 S.siz(4)/2-100],'Units','pixels');
+        set(S.panel2a,'Position',[210 120 2000 S.siz(4)/2-100],'Units','pixels');
+
+        S.slider2=uicontrol('Style','Slider','Parent',S.fh,'Units','pixels','Position',[210 120 S.siz(3)-220 0.05],'Value',0.5,'Callback',@slider_callback2);
+        
+        S.ax2 = axes('OuterPosition',[0.02 0.02 0.98 0.98],'Units','normalized','Parent',S.panel2a);
         % make initial plot
[…]
