### 2024-01-20 ff8f661
adjusted ReadPlotLayerPicks for new layer picking flags (line or points)
