### 2024-01-26 25186de
new figure for LayerPicking
