+++ Scripts_for_3D/Mala3D.m
@@ -19,10 +19,10 @@ clc
 platform=2; % Linux=1, Mac=2, Windows=3
 
 % Select number of profiles:
-profile_min=1;  % minimum profile number
-profile_max=90;  % maximum profile number
+profile_min=3;  % minimum profile number
+profile_max=6;  % maximum profile number
 % number of channels for this dataset
-channels=16; % number of channels
+channels=22; % number of channels
 
 changeDir=0; % if =1: change the sign of the y-antenna-GPS-offset, if =0: use offsets as written in file
 
@@ -51,7 +51,7 @@ continuerectangle=5;    % number of starting rectangle
 %%% Raw data is only read if userawdata=0! Otherwise mat-files of raw
 %%% data in folder profiles2mat are used.
 userawdata=0;  % if =1: use aready read in raw data and apply new processing steps
-useprocdata=1;  % if =1: use already processed data in folder profiles2mat/proc for binning
+useprocdata=0;  % if =1: use already processed data in folder profiles2mat/proc for binning
 
 % suppress display of figures
 nofig=0;    % if =1: no figures will be shown
