### 2024-01-26 4070710
Update paper.bib (corrected and added doi)
