+++ paper/paper.bib
@@ -31,6 +31,7 @@ doi = {10.1144/jgs2023-042}
   number={3},
   pages={781},
   year={2022},
+  doi={10.3390/rs14030781},
   publisher={MDPI}
 }
 
@@ -42,6 +43,7 @@ doi = {10.1144/jgs2023-042}
   number={14},
   pages={3647},
   year={2023},
+  doi={10.3390/rs15143647},
   publisher={MDPI}
 }
 
@@ -52,7 +54,7 @@ journal = {Soil Use and Management},
 volume = {39},
 number = {4},
 pages = {1289-1303},
-doi = {https://doi.org/10.1111/sum.12964},
+doi = {10.1111/sum.12964},
 year = {2023}
 }
 
@@ -63,6 +65,7 @@ year = {2023}
   journal={Proceedings of the 17th International Conference on Ground Penetrating Radar, Rapperswil, Switzerland, 18–21 June 2018},
   year={2018},
   pages={1-5},
+  doi={10.1109/ICGPR.2018.8441658},
   url={https://emanuelhuber.github.io/RGPR/2018_huber-and-hans_RGPR-new-R-package_notes.pdf}
 }
 
@@ -85,7 +88,7 @@ publisher = {Elsevier},
 pages = {341-398},
 year = {2019},
 isbn = {978-0-12-812429-1},
-doi = {https://doi.org/10.1016/B978-0-12-812429-1.00010-6},
+doi = {10.1016/B978-0-12-812429-1.00010-6},
 url = {https://www.sciencedirect.com/science/article/pii/B9780128124291000106},
 author = {{Pajewski}, L. and {Fontul}, S. and {Solla}, M.},
 }
@@ -109,6 +112,7 @@ author = {{Pajewski}, L. and {Fontul}, S. and {Solla}, M.},
   journal={Proceedings of the XIII International Conference on Ground Penetrating Radar},
   year={2010},
   pages={1-5},
+  doi={10.1109/ICGPR.2010.5550131},
   url={https://api.semanticscholar.org/CorpusID:11558802}
 }
 
@@ -120,6 +124,7 @@ author = {{Pajewski}, L. and {Fontul}, S. and {Solla}, M.},
   number={3},
   pages={171-195},
   year={2018},
+  doi={10.1002/arp.1599},
   publisher={Wiley Online Library}
 }
 
@@ -131,6 +136,7 @@ author = {{Pajewski}, L. and {Fontul}, S. and {Solla}, M.},
   number = {1},
   year={2010},
   pages={17-43},
+  doi={10.1109/ICGPR.2010.5550131},
   url={https://api.semanticscholar.org/CorpusID:11558802}
 }
 
@@ -154,7 +160,7 @@ DOI = {10.5194/tc-4-475-2010}
   volume = {45},
   number = {1},
   pages={279-283},
-  doi={https://doi.org/10.4000/archeosciences.10100}
+  doi={10.4000/archeosciences.10100}
 }
 
 @article{Wunderlich:2022,
@@ -165,6 +171,7 @@ DOI = {10.5194/tc-4-475-2010}
   number={15},
   pages={3665},
   year={2022},
+  doi={10.3390/rs14153665},
   publisher={MDPI}
 }
 
@@ -173,6 +180,7 @@ DOI = {10.5194/tc-4-475-2010}
   author={{Wunderlich}, T. and {Brozio}, J.P. and {Feeser}, I. and {Heum{\"u}ller}, M. and {Mohr}, C.},
   booktitle={Advances in On-and Offshore Archaeological Prospection: Proceedings of the 15th International Conference on Archaeological Prospection},
[…]
