### 2024-01-26 #6248 @tinawunderlich
Now I included the changes also in the paper and fixed/added the dois there.
