### 2024-01-29 #6248 @tinawunderlich
Hi @kthyng , as I had the software before in our universersity-internal GitLab, I exported the log of my activity since the starting point (July 2019). Do you want me to comment it for you to make it clearer or do you just want to have it as it is? And shall I post it here as attachment or send it via email? Thanks!
