### 2024-02-13 #6248 @tinawunderlich
@kthyng I attach the log of the former gitlab-account to this post. I marked the changes in different colours in order make it easier for the reader to know which line belongs to which sort of change. I hope that helps!
[MultichannelGPR_Log.pdf](https://github.com/openjournals/joss-reviews/files/14271649/MultichannelGPR_Log.pdf)
