### 2024-02-27 8038846
(1) increased padding width in topomig2d to account for very steep topography; (2) added output of recommended zmin and zmax for the migration in prepareMigrationFiles
