### 2024-02-27 813e530
corrected readdzt_4000 for measurements west of Greenwich -> negative longitude for utm-conversion
