### 2024-03-11 #6248 @tinawunderlich
Hi @kthyng, thanks for your comments! To be honest, I've not thought about making it a Matlab toolbox before. But now I had a look at it. I think that it will be very nice for a "finished" program, but in the current state of my program, when I'm stil going to work on it on a daily/weekly basis, this would be not a good option. The same applies to the Matlab documentation, which I understood is linked to a toolbox and cannot be created for a "collection of scripts" as my program is now. So my opinion on this is that I will keep it in mind for the future, when the program is going to be more static.
Regarding the unit tests: I think this is a good option and I will try to work on that, but it will take some time for me to learn about and integrate it.
The rephrasing of the title is a good suggestion, especially as JOSS is not a solely geophysical journal. I'll change the title and upload the paper again.
