+++ Processing/interpolation.m
@@ -4,7 +4,7 @@ function [traces]=interpolation(traces,gap)
 %
 % [traces]=interpolation(traces,gap)
 %
-% Dr. Tina Wunderlich, CAU Kiel 2020, tina.wunderlich@ifg.uni-kiel.de
+% Dr. Tina Wunderlich, CAU Kiel 2020/2024, tina.wunderlich@ifg.uni-kiel.de
 %
 % Input:
 % traces: Matrix with traces along profile (traces in columns), missing
@@ -14,16 +14,25 @@ function [traces]=interpolation(traces,gap)
 % Output:
 % traces: Matrix with traces using mean of neighboring traces
 
+x=find(~isnan(traces(1,:))); % ~isnan in first sample row = value
+trnum=1:length(traces(1,:)); % all trace numbers
+miss=find(isnan(traces(1,:))); % number of missing traces
 
+% interpolate everything
+ns=length(traces(:,1));
+for i=1:ns % for each sample
+    traces(i,:)=interp1(x,traces(i,x),trnum,'linear','extrap');
+end
 
-
-for i=2:length(traces(1,:)) % for all traces
-    a=1;
-    while a<=gap && i+a<length(traces(1,:))
-        if ~isnan(traces(1,i-1)) && isnan(traces(1,i)) && ~isnan(traces(1,i+a))
-            traces(:,i:i+a-1)=repmat(mean(traces(:,[i-1 i+a]),2),[1 a]);
-            break;
-        end
-        a=a+1;
+% create mask taking into account the gap
+mask=NaN(size(trnum));
+mask(x)=1; % set given traces to valid
+for i=1:length(miss)
+    if min(abs(miss(i)-x))<gap
+        mask(miss(i))=1;
     end
 end
+
+% apply mask
+mask=repmat(mask,[ns,1]);
+traces=traces.*mask;
