+++ Scripts_for_3D/Radargrams2bins.m
@@ -21,7 +21,7 @@ clc
 platform=2; % Linux=1, Mac=2, Windows=3
 
 % Bin size of grid
-dx=0.04; % [m] 
+dx=0.08; % [m] 
 
 % Automatic rotation of measurement area for minimum memory size
 rotate_area=1;  % 1=yes, 0=no
@@ -29,7 +29,7 @@ rotate_area=1;  % 1=yes, 0=no
 % Division into rectangles for processing
 border=0.5;   % overlapping border of rectangles in m
 
-force_division=1; % if =1: force program to divide into several rectangles
+force_division=0; % if =1: force program to divide into several rectangles
 
 % Parallelization (set this to 0 if you run into memory problems or if you
 % do not have a parallel pool)
