+++ Export_Import/convert_PPS2pos.m
@@ -16,17 +16,23 @@ clc
 %%-------------------------------------------------------------------------
 
 
-filter_coords=1; % if =1: coordinates are filtered and traces at
+filter_coords=2; % possible values: 0, 1, 2:
+% if =0: no filtering
+% if =1: coordinates are filtered and traces at
 % approx. the same position will be removed (e.g. at the beginning or
 % end of the profile): a local difference
 % over m coordinates for x and y is calculated separately. If this
 % absolute local difference is smaller than nstd*std(diff) it will be removed.
-m=15; % number of samples for local difference calculation
-nstd=1.5; % nstd*standard deviation as valid range
-show_filter_plot=0; % if =1: show plot for quality control of filtering
+% if =2: coordinates are filtered: local differences over m coordinates are
+% calculated for x and y separately and points with differences around zero
+% will be deleted (traces at the same place) using nstd as a threshold in
+% meters (-/+)
+m=9; % number of samples for local difference calculation (must be odd!)
+nstd=0.02; % if filter_coords==1: nstd*standard deviation as valid range, if filter_coords==2: +/-nstd [m] will be deleted
+show_filter_plot=1; % if =1: show plot for quality control of filtering
 
-coords_smooth=1; % if=1: smooth coordinates along profile
-nsmooth=15; % number of traces for smoothing of coords
+coords_smooth=1; % if=1: smooth coordinates along profile (will be applied after filtering)
+nsmooth=5; % number of traces for smoothing of coords
 
 
 %% DO NOT CHANGE FROM HERE ON
@@ -75,6 +81,8 @@ end
 
 % get all available *.pos-files with local coords
 list=dir([foldername,'/*.pos']);
+weg=cellfun(@(x) startsWith(x,'.'),{list.name}); % these are just temporary files...
+list(weg)=[]; % ... delete them from list
 
 if ~isempty(list)
     disp([int2str(length(list)),' *.pos-files found. Moving to folder local_coords.'])
@@ -83,33 +91,61 @@ if ~isempty(list)
         mkdir(fullfile(foldername,'local_coords'));
     end
     for i=1:length(list)
-        if ~startsWith(list(i).name,'.')
-            % get profile number
-            test=extractBetween(list(i).name,'_','_');
-            number(i,1)=str2num(test{1});
+        % get profile number
+        test=extractBetween(list(i).name,'_','_');
+        number(i,1)=str2num(test{1});
 
-            disp(int2str(number(i,1)))
+        disp(['   Profile ',int2str(number(i,1))])
 
-            % read utm coords
-            temp=extractBefore(list(i).name,'_');
-            utm=load(fullfile(foldername,'..',[temp,'_',test{1},'.utm'])); % trace number, N, E, height, UTMZone
-            utm(utm(:,1)==-1,:)=[];
+        % read utm coords
+        temp=extractBefore(list(i).name,'_');
+        utm=load(fullfile(foldername,'..',[temp,'_',test{1},'.utm'])); % trace number, N, E, height, UTMZone
+        utm(utm(:,1)==-1,:)=[];
 
-            % move original file to new folder
-            movefile(fullfile(foldername,list(i).name),fullfile(foldername,'local_coords',list(i).name));
+        % move original file to new folder
+        movefile(fullfile(foldername,list(i).name),fullfile(foldername,'local_coords',list(i).name));
 
-            % determine time offset with PPS and apply it
-            utm=PPS_synchro(foldername,list(i).name,utm);
[…]
