+++ DZT_Import/DZT_Convert.m
@@ -11,15 +11,15 @@ clear all
 close all
 clc
 
-app='SIR4000';    % Equipment: SIR20 / SIR30 / SIR3000 / SIR4000 / Tablet
+app='Tablet';    % Equipment: SIR20 / SIR30 / SIR3000 / SIR4000 / Tablet
 
 dataplot=1; % plot radargram for controlling? 1=yes, 0=no
 
 convert2utm=1; % convert WGS Lat/Long to UTM (=1 if measured with Stonex-GPS)
-zone=29; % if convert2utm==1 -> give UTM-zone
+zone=32; % if convert2utm==1 -> give UTM-zone
 
 offsetGPS_X=0; % Offset between GPS and antenna midpoint crossline (in profile direction GPS left of antenna -> positive)
-offsetGPS_Y=0; % Offset between GPS and antenna midpoint in profile direction (if GPS behind antenna midpoint -> positive)
+offsetGPS_Y=-0.12; % Offset between GPS and antenna midpoint in profile direction (if GPS behind antenna midpoint -> positive)
 
 % Options for calculating inline coordinates for each trace:
 coords_opt=1;   % =1: trace coordinate is difference to beginning of profile (only use this for straight profiles!)

[…]
