### 2024-03-12 1ba07b2
added option for median filter in Process_SingleProfile for mat-files
