### 2024-03-12 6318c4d
(1) added trace-wise median filter in processing.m; (2) finally corrected the colorscale bug in Process_SingleProfile for mat-files
