### 2024-03-13 a5b4072
(1) removed otpion for GSSI data in Process_SingleProfile as it can be done with mat-files; (2) updated the options for Mala-processing in Process_SingleProfile by including constant trace distance, interpolation and trace-wise median filtering; (3) updated Mala3D to have the same processing options as Process_SingleProfile; (4) corrected the colorscale bug in Process_SignleProfile for Mala-data
