+++ GUIs/processingTestGUI.m
@@ -135,7 +135,7 @@ S.fh.Visible='on';
         S.tendtext=uicontrol(S.fh,'Style','text','String','tend [ns]','Position',[360-off1 240 140 20],'HorizontalAlignment','left');
         
         % Trace interpolation
-        S.traceinterp=uicontrol(S.fh,'Style','checkbox','String','Trace interpolation','FontWeight','bold','Position',[300-off1 210 200 15],'Value',0,'Callback',@traceinterp_call);
+        S.badtrace=uicontrol(S.fh,'Style','checkbox','String','Bad trace removal','FontWeight','bold','Position',[300-off1 210 200 15],'Value',0,'Callback',@badtraceremoval_call);
         S.minfactor=uicontrol(S.fh,'Style','edit','String','3','Enable','off','Position',[320-off1 180 35 20]);
         S.maxfactor=uicontrol(S.fh,'Style','edit','String','3','Enable','off','Position',[320-off1 150 35 20]);
         S.minfactortext=uicontrol(S.fh,'Style','text','String','minfactor','Position',[360-off1 180 140 20],'HorizontalAlignment','left');
@@ -245,7 +245,7 @@ S.fh.Visible='on';
         S.aperture2text=uicontrol(S.fh,'Style','text','String','aperture [°]','Position',[1160-off1-off5 90 140 20],'HorizontalAlignment','left');
         
         % 6th column:
-        off6=130;
+        off6=120;
         % spectralWhitening
         S.sw=uicontrol(S.fh,'Style','checkbox','String','Spectral Whitening','FontWeight','bold','Position',[1300-off1-off6 300 200 15],'Value',0,'Callback',@sw_call);
         S.fmin_sw=uicontrol(S.fh,'Style','edit','String','100','Enable','off','Position',[1320-off1-off6 270 35 20]);
@@ -254,6 +254,22 @@ S.fh.Visible='on';
         S.fmax_sw_text=uicontrol(S.fh,'Style','text','String','fmax [MHz]','Position',[1360-off1-off6 240 140 20],'HorizontalAlignment','left');
         S.alpha=uicontrol(S.fh,'Style','edit','String','0.01','Enable','off','Position',[1320-off1-off6 210 35 20]);
         S.alpha_text=uicontrol(S.fh,'Style','text','String','alpha','Position',[1360-off1-off6 210 140 20],'HorizontalAlignment','left');
+
+        off6=-80;
+        % median filter on traces
+        S.medfilt=uicontrol(S.fh,'Style','checkbox','String','Trace-wise median filter','FontWeight','bold','Position',[1100-off1-off6 180 200 15],'Value',0,'Callback',@medfilt_call);
+        S.numsamp=uicontrol(S.fh,'Style','edit','String','3','Enable','off','Position',[1120-off1-off6 150 35 20]);
+        S.numsamptext=uicontrol(S.fh,'Style','text','String','number of samples','Position',[1160-off1-off6 150 140 20],'HorizontalAlignment','left');
+
+        % Const Trace Distance
+        S.constDist=uicontrol(S.fh,'Style','checkbox','String','Constant trace distance','FontWeight','bold','Position',[1100-off1-off6 120 200 15],'Value',0,'Callback',@constDist_call);
+        S.dist=uicontrol(S.fh,'Style','edit','String','0.02','Value',0,'Enable','off','Position',[1120-off1-off6 90 35 20]);
+        S.disttext=uicontrol(S.fh,'Style','text','String','dx [m]','Position',[1160-off1-off6 90 140 20],'HorizontalAlignment','left');
+
+        % Trace interpolation
[…]
