+++ GUIs/processingTestGUI_gssi.m
@@ -1,1162 +1,0 @@
-function []=processingTestGUI_gssi(folder,name,profilelist,UTMzone)
-
-% GUI for plotting of profiles and interactive testing of processing steps
-%
-% Dr. Tina Wunderlich, CAU Kiel 2020, tina.wunderlich@ifg.uni-kiel.de
-%
-% Input:
-% folder: path and name of rSlicer-folder
-% name: name of data
-% profilelist: number of profiles
-
-
-S.folder=folder;
-S.name=name;
-S.profilelist=profilelist;
-
-% load first profile
-%[S.traces,S.dt,S.ns,S.x,S.y,S.z]=readmala(S.folder,S.name,S.profilelist(1));
-[S.traces,S.dt,S.ns,S.x,S.y,S.z]=readGSSI(S.folder,UTMzone,S.profilelist(1));
-
-
-S.channellist=[{'All'}];
-S.raw=[];
-S.info=[]; % channel number corresponding to traces in raw/data
-S.zz=[];
-for i=1:length(S.traces)
-    S.channellist=[S.channellist; {int2str(i)}];
-    S.raw=[S.raw S.traces{i}];
-    S.xprof{i}=[0; cumsum(sqrt(diff(S.x{i}).^2+diff(S.y{i}).^2))];
-    S.info=[S.info zeros(1,length(S.xprof{i}))+i];
-    S.zz=[S.zz; S.z{i}];
-end
-S.xloc=[S.xprof{1}];
-for i=2:length(S.traces)
-    S.xloc=[S.xloc; S.xprof{i}+max(S.xloc)];
-end
-S.t=0:S.dt:S.dt*(S.ns-1);
-
-% traces are original raw data in cells, raw ist raw data of all channels, proc
-% is processed data of all channels, choose channels via info (channel
-% number for each trace), xloc is x-axis for all channels, xprof is x-axsi
-% for each channel in cells
-
-% get screensize
-S.siz = get( 0, 'Screensize' );
-S.fh.Position=S.siz;  % default: screensize
-
-% Add the UI components
-S=addcomponents(S);
-
-% Make figure visible after adding components
-S.fh.Visible='on';
-
-    function S=addcomponents(S)
-        
-        % create figure handle:
-        S.fh = figure('units','pixels',...
-            'position',S.siz,...
-            'menubar','none',...
-            'name','Processing test',...
-            'numbertitle','off',...
-            'resize','on','Visible','off','SizeChangedFcn',@resizeui);
-        
-        % make initial plot of radargram
-        S.ax=axes('unit','pix',...
-            'position',[250 300 S.siz(3)-300 S.siz(4)-320]);
-        S.rad=imagesc(S.xloc,S.t,S.raw);
-        hold on
-        for i=1:length(S.xprof)-1
-            plot([max(S.xloc(S.info==i)) max(S.xloc(S.info==i))],[0 max(S.t)],'r')
-        end
-        colormap(flipud(gray));
-        xlabel('x [m]')
-        ylabel('t [ns]')
-        set(S.ax,'DataAspectratio',[1 1 1])
-        grid on
-        axis tight
-        
-        %% UI controls:
-        % UI control for channel number
-        S.chan=uicontrol(S.fh,'style','listbox','unit','pix','position',[20 310 55 100],'String',S.channellist,'Value',1,'callback',@channum_call);
-        S.chantext=uicontrol(S.fh,'style','text','unit','pix','position',[20 420 55 15],'String','Channel','HorizontalAlignment','left');
-        
-        % UI control for profile number
-        S.prof=uicontrol(S.fh,'Style','listbox','unit','pix','position',[80 310 55 100],'String',S.profilelist,'Value',1,'callback',@profnum_call);
[…]
