### 2024-03-13 a642b4f
added option for plotting picks in figures (not for georeferenced figures!) in saveFigures.m
