### 2024-03-13 7ee9c90
Velocity2D: (1) deleted the semblance part (was not working properly) and instead incorporated the automatic v determination like the best method in Wunderlich et al. 2022 (Hyperbola detection...); (2) improved the program by automatically deleting not-saved hyperbola and showing saved hyperbola again when the profile/channel is changed
