+++ GUIs/Velocity_2D_picking_plot.m
@@ -57,6 +57,8 @@ S.cmax1=coldata(end-round(length(coldata)/100*1));
 S.cmin3=coldata(round(length(coldata)/100*3));
 S.cmax3=coldata(end-round(length(coldata)/100*3));
 
+S.tbox=10; % default length of t for box for auto v in ns
+
 % get screensize
 S.siz=get(0,'ScreenSize');
 S.fh.Position=S.siz;  % default: screensize
@@ -97,13 +99,13 @@ uimenu(figureToolBar,'Label','Pan','Callback','pan on');
         S.asptext=uicontrol(S.fh,'style','text','unit','pix','position',[20 85 100 20],'String','Aspect ratio t','HorizontalAlignment','left');
         
         % UI control - radiobuttons for colorscale
-        S.rb1=uicontrol('Style','radiobutton','String','Auto color scale','Position',[280 60 150 15],'Value',1,'Callback',@colorscale_call); % this button is on
+        S.rb1=uicontrol('Style','radiobutton','String','Auto color scale','Position',[280 60 150 15],'Callback',@colorscale_call); 
         S.rb2=uicontrol('Style','radiobutton','String','1 %','Position',[280 40 150 15],'Callback',@colorscale_call);
-        S.rb3=uicontrol('Style','radiobutton','String','3 %','Position',[280 20 150 15],'Callback',@colorscale_call);
-        S.flag_cs=1;
+        S.rb3=uicontrol('Style','radiobutton','String','3 %','Position',[280 20 150 15],'Value',1,'Callback',@colorscale_call); % this button is on
+        S.flag_cs=3;
         
         % UI control: gain
-        S.gain_cb=uicontrol(S.fh,'Style','checkbox','unit','pix','Position',[150 120 100 15],'String','Apply gain [dB]','Value',1,'Callback',@gain_call);
+        S.gain_cb=uicontrol(S.fh,'Style','checkbox','unit','pix','Position',[150 120 100 15],'String','Apply gain [dB]','Value',0,'Callback',@gain_call);
         S.gain1 = uicontrol(S.fh,'Style','edit','unit','pix','Position',[150 100 60 15],'String','-20','Callback',@gain_call);
         S.gain2 = uicontrol(S.fh,'Style','edit','unit','pix','Position',[150 80 60 15],'String','0','Callback',@gain_call);
         S.gain3 = uicontrol(S.fh,'Style','edit','unit','pix','Position',[150 60 60 15],'String','15','Callback',@gain_call);
@@ -114,16 +116,20 @@ uimenu(figureToolBar,'Label','Pan','Callback','pan on');
         S.new=uicontrol(S.fh,'Style','pushbutton','unit','pix','position',[480 55 100 20],'callback',{@new_call},'String','Make new v-pick');
         
         % UI control for semblance button
-        S.semb=uicontrol(S.fh,'Style','checkbox','unit','pix','position',[480 20 150 20],'callback',{@semb_call},'Value',0,'String','Auto v (via semblance)');
+        S.semb=uicontrol(S.fh,'Style','checkbox','unit','pix','position',[480 20 150 20],'callback',{@semb_call},'Value',0,'String','Auto v');
         
         % UI control for v-slider
-        S.vSlider=uicontrol(S.fh,'Style','slider','unit','pix','position',[650 30 100 20],'callback',{@v_call},'min',3,'max',18,'Value',10,'sliderstep',[0.1/15 1/15],'Enable','off');
-        S.text6=uicontrol(S.fh,'style','text','unit','pix','position',[650 55 100 20],'String','v =           cm/ns');
-        S.vtext=uicontrol(S.fh,'style','text','unit','pix','position',[675 55 30 20],'String','10');
[…]
