+++ GUIs/Velocity_2D_picking_plot.m
@@ -81,8 +81,28 @@ uimenu(figureToolBar,'Label','Pan','Callback','pan on');
         S.fh= figure('menubar','none','Position',S.siz,...
             'name','Velocity picking',...
             'numbertitle','off',...
-            'resize','on','Visible','off','SizeChangedFcn',@resizeui);
+            'resize','on','Visible','off','SizeChangedFcn',@resizeui,'Renderer','painters');
         
+        % scrollbar for plot:
+        S.panel1=uipanel('Parent',S.fh);
+        S.panel1a=uipanel('Parent',S.panel1);
+        set(S.panel1,'Position',[150 160 S.siz(3)-100 S.siz(4)-200],'Units','pixels'); % panel for cutting frame with slider
+        set(S.panel1a,'Position',[150 160+S.siz(3)/2-100 S.siz(3) S.siz(4)-200],'Units','pixels'); % panel inside cutting frame with image
+
+        S.slider1=uicontrol('Style','Slider','Parent',S.fh,'Units','pixels','Position',[150 160 S.siz(3)-100 0.05],'Value',0.5,'Callback',@slider_callback1);
+        
+        % Plot of radargram
+        S.ax = axes('units','normalized','Parent',S.panel1a,...
+            'OuterPosition',[0.02 0.02 0.98 0.98]);
+        S.Plot = imagesc(S.ax,S.coord,S.t,S.traces);
+        hold on
+        xlabel('x [m]')
+        ylabel('t [ns]')
+        grid on
+        title(['Profile ',int2str(S.profiles(1)),', Channel ',int2str(S.currchan)])
+        colormap(flipud(gray))
+        set(S.ax,'DataAspectratio',[1/20 1 1],'CLim',[S.cmin3 S.cmax3])
+
         %%% UI control elements:
         % UI control - listbox for channel number
         S.chan = uicontrol(S.fh,'style','listbox','unit','pix','position',[20 220 50 60],'callback',{@chan_call},'Value',1,'String',int2str([1:S.chan]'));
@@ -132,23 +152,13 @@ uimenu(figureToolBar,'Label','Pan','Callback','pan on');
         S.text7=uicontrol(S.fh,'style','text','unit','pix','position',[880 20 160 15],'String','Half width of hyperbola [m]','HorizontalAlignment','Left');
         
         % UI control for saving of velocities
-        S.save=uicontrol(S.fh,'Style','pushbutton','unit','pix','position',[800 55 120 20],'callback',{@save_call},'String','Save current velocity','Enable','off');
+        S.save=uicontrol(S.fh,'Style','pushbutton','unit','pix','position',[830 55 120 20],'callback',{@save_call},'String','Save current velocity','Enable','off');
         
         % UI control for tmax
         S.tmax=uicontrol(S.fh,'Style','edit','unit','pix','position',[20 140 50 20],'callback',{@tmax_call},'Value',max(S.t),'String',num2str(max(S.t)));
         S.text8=uicontrol(S.fh,'style','text','unit','pix','position',[20 160 80 20],'String','tmax [ns]','HorizontalAlignment','Left');
         
-        % Plot of radargram
-        S.ax = axes('unit','pix',...
-            'position',[200 140 S.siz(3)-300 S.siz(4)-200]);
-        S.Plot = imagesc(S.coord,S.t,S.traces);
-        hold on
-        xlabel('x [m]')
-        ylabel('t [ns]')
-        grid on
-        title(['Profile ',int2str(S.profiles(1)),', Channel ',int2str(S.currchan)])
-        colormap(flipud(gray))
-        set(S.ax,'DataAspectratio',[S.aspval 1 1],'CLim',[S.cmin3 S.cmax3])
+        
         
         % read pickfile:
[…]
