### 2024-03-14 590c968
(1) improved output of make_Timeslices when followTopo is on (this flag is saved in depth.mat); (2) plot_Timeslices: corrected title of Tsl is followTopo is used; (3) improved Tsl_Radargrams: scrollbars on radargram and timeslices, zoom for timeslices, wiggle-option (for a very small amount of traces only!) a little faster, plotting of timeslices-range in radargram, also for followTopo
