+++ GUIs/plot_TslRadargram.m
@@ -1,4 +1,4 @@
-function []=plot_TslRadargramm(tsl,xgrid,ygrid,t_tsl,radar,t,coords,profileslist,timedepth)
+function []=plot_TslRadargramm(tsl,xgrid,ygrid,t_tsl,radar,t,coords,profileslist,timedepth,pfad_tsl,maxElevation,followTopo)
 
 % GUI for plotting of Tsl and radargrams with connected cursor
 %
@@ -13,12 +13,20 @@ function []=plot_TslRadargramm(tsl,xgrid,ygrid,t_tsl,radar,t,coords,profileslist
 % radar: cells with data of radargrams 
 % t: time vector of radargrams [ns]
 % coords: cells with starting and ending coordinates of profile (or
-% complete profile lists for curved profiles)
+% complete profile lists for curved profiles)=global_coords
 % profilelist: number of profiles
 % timedepth: flag: time=1, depth=2
+% pfad_tsl: path to tsl folder
+% maxElevation: maximum elevation for depthslices in m
+% followTopo: if=1 slices following topography, if=0 horizontal slices
+% topo: cell array for each radargram with topography
+
 
 
 S.timedepth=timedepth;
+S.maxElevation=maxElevation;
+S.followTopo=followTopo;
+S.pfad_tsl=pfad_tsl; % path to tsl folder (for coordtrans)
 % set tsl
 S.x = xgrid;
 S.y = ygrid;
@@ -26,14 +34,18 @@ S.t_tsl = t_tsl;
 S.tsl=tsl;
 S.dx=xgrid(1,2)-xgrid(1,1);
 S.tslnum=1; % current number of tsl
-S.coordtrans=[];
-S.ar=1; %aspect ratio factor
+S.ar=1/10; % aspect ratio factor
 
 % set radargramms (=data)
 S.data=radar;
 S.t=t;
 if timedepth==2
     S.t=abs(S.t-S.t(1)); % depth is positive down, starting at 0m -> corresponds to dsl
+    S.t=S.maxElevation-S.t; % absolute heights
+
+    % also for dsl:
+    S.t_tsl=S.maxElevation-S.t_tsl;
+    S.t_tsl=fliplr(S.t_tsl); % to make first row smaller than second
 end
 S.dt=S.t(2)-S.t(1);
 for i=1:length(coords)
@@ -47,16 +59,68 @@ for i=1:length(coords)
     S.data{i}=S.data{i}(:,a);
 end
 S.profileslist=profileslist;
-S.rdgnum='1';
+S.rdgnum='1'; % radargram number
+
+S.factor=1; % zoom factor for tsl plot
+
+if S.followTopo==1
+    % prepare red lines for slices following topography:
+    for i=1:length(coords) % for each profile
+        S.lines{i}=repmat(coords{i}(:,3)',[length(S.tsl)+1,1])-repmat([0; cumsum(abs(S.t_tsl(:,1)-S.t_tsl(:,2)))],[1,length(coords{i}(:,3))]); % one row of topo for each tsl-border
+    end
+end
+
+
+if exist(fullfile(S.pfad_tsl,'coordtrans.mat'),'file')
+    S.coordtrans=load(fullfile(pfad_tsl,'coordtrans.mat')); % lokal/global
+else
+    % get path to coordtrans:
+    if ~ispc; menu('Choose folder with coordtrans.mat (e.g. 3D_Grid_R*)','OK'); end
+    if exist('.cttemp.temp') % read last opened folder from temp.temp
+        fid=fopen('.cttemp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            pfad_ct=uigetdir(fn{1}{1},'Choose folder with coordtrans.mat (e.g. 3D_Grid_R*)');
+        else
[…]
