### 2024-03-16 #6248 @thomakra
Apologies, but now is not a good time to contribute.

On Thu, Mar 14, 2024, 15:28 Martin Fleischmann ***@***.***>
wrote:

> Hi @LyceanEM <https://github.com/LyceanEM> and @thomakra
> <https://github.com/thomakra>, would you be willing to help JOSS by
> reviewing this submission? Thank you!
>
