### 2024-03-20 ea2dece
corrected bug in TurnProfiles
