### 2024-03-22 d9fb128
added GUI for plotting Radargrams and corresponding map (for sparse profiles or if no timeslices are available)
