+++ GUIs/plot_MapRadargram.m
@@ -1,0 +1,508 @@
+function []=plot_MapRadargramm(radar,t,coords,profileslist,timedepth)
+
+% GUI for plotting of map and radargrams with connected cursor
+%
+% Dr. Tina Wunderlich, CAU Kiel 2024, tina.wunderlich@ifg.uni-kiel.de
+%
+% Input:
+% (created by make_Radargrams.m, Bin2Radargram.m or own creation)
+% radar: cells with data of radargrams 
+% t: time vector of radargrams [ns]
+% coords: cells with starting and ending coordinates of profile (or
+% complete profile lists for curved profiles)=global_coords
+% profilelist: number of profiles
+% timedepth: flag: time=1, depth=2
+
+
+S.timedepth=timedepth;
+
+% set radargramms (=data)
+S.data=radar;
+S.t=t;
+S.maxElevation=S.t(1);
+S.dt=abs(S.t(2)-S.t(1));
+for i=1:length(coords)
+    S.xprof{i}=coords{i}(:,1);  % x-coordinate along profile
+    S.yprof{i}=coords{i}(:,2);  % y-coordinate along profile
+    S.loc_coord{i}=[0; cumsum(sqrt(diff(S.xprof{i}).^2+diff(S.yprof{i}).^2))];  % local coordinate along profile
+    % delete possibly same coordinates:
+    [S.loc_coord{i},a]=unique(S.loc_coord{i});
+    S.xprof{i}=S.xprof{i}(a);
+    S.yprof{i}=S.yprof{i}(a);
+    S.data{i}=S.data{i}(:,a);
+end
+S.profileslist=profileslist;
+S.rdgnum='1'; % radargram number
+S.ar=1/10;
+
+
+% colorscale limits radargrams
+for i=1:length(S.data)
+    coldata=sort(unique(S.data{i}(~isnan(S.data{i}(:)))));
+    S.cmin1R(i)=coldata(round(length(coldata)/100*1));
+    S.cmax1R(i)=coldata(end-round(length(coldata)/100*1));
+    S.cmin3R(i)=coldata(round(length(coldata)/100*3));
+    S.cmax3R(i)=coldata(end-round(length(coldata)/100*3));
+end
+
+% get screensize
+S.siz = get( 0, 'Screensize' );
+S.fh.Position=S.siz;  % default: screensize
+
+% Add the UI components
+S=addcomponents(S);
+
+% Make figure visible after adding components
+S.fh.Visible='on';
+
+    function S=addcomponents(S)
+        
+        % create figure handle: (Renderer is important for scrollbar)
+        S.fh = figure('units','pixels',...
+            'position',S.siz,...
+            'menubar','none',...
+            'name','Map and Radargrams',...
+            'numbertitle','off',...
+            'resize','on','Visible','off','SizeChangedFcn',@resizeui,'Renderer','painters','KeyPressFcn',@keypress);
+        
+        S.fh.MenuBar = 'figure';
+        set (S.fh, 'WindowButtonMotionFcn', @mouseMove); % for connecting of Tsl and radargrams
+        
+        %%% RADARGRAM
+        % scrollbar for radargram:
+        S.panelR=uipanel('Parent',S.fh);
+        S.panelRa=uipanel('Parent',S.panelR);
+        if range(S.yprof{1})>range(S.xprof{1}) % Map height>width
+            % layout1
+            set(S.panelR,'Position',[180+(S.siz(3)-200)/3 40 S.siz(3)-200-(S.siz(3)-200)/3 S.siz(4)/3],'Units','pixels'); % panel for cutting frame with slider
[…]
