+++ Visualization/View_3Dblock.m
@@ -14,7 +14,7 @@ clc
 
 
 % Number of rectangles with binned data
-rectangles=1; % e.g. 1:3
+rectangles=1:2; % e.g. 1:3
 
 % depth or time domain?
 dsl=0;  % =1: depth, =0: time
