### 2024-03-26 e9876dc
prepareMigrationFiles: added option for reading in a file with several topography points on the measurement area, which is then interpolated and teh topography is extracted at the radargram locations
