+++ Scripts_for_2D/prepareMigrationFiles.m
@@ -5,7 +5,7 @@ clc
 
 % Prepare topo.mat and vgrid.mat for Migrations in Processing.m
 %
-% Dr. Tina Wunderlich, CAU Kiel 2021, tina.wunderlich@ifg.uni-kiel.de
+% Dr. Tina Wunderlich, CAU Kiel 2021-2024, tina.wunderlich@ifg.uni-kiel.de
 %
 % requires folder with radargrams.mat/global_coords.mat/x.mat/t.mat
 % -> choose this folder when asked!
@@ -18,19 +18,21 @@ vconst=0.08; % v in m/ns
 % vopt=2: Constant velocity (but different) for each profile in radargrams.mat
 vall=[0.1 0.08]; % v for each profile in m/ns
 % vopt=3: 1D velocity model for all profiles
-v1d=[0.16 0.1]; % vRMS at different times in m/ns
+v1d=[0.16 0.1]; % v at different times in m/ns
 t1d=[0 18]; % corresponding times in ns
 
 % Choose options for topography:
-topoopt=2;
+topoopt=3;
 % topoopt=0: no topography file required
 % topoopt=1: topography is already set in global_coords(:,3)
 % topoopt=2: topography has to be set with file containing profile number (1. column),
 % profile coordinate (2. Column) and topography (3. column), all in m, This
-% file should be in the same folder as radargrams.mat. The smoothed
-% topography is then also set in global_coords as height.
-topofile='Topo.txt'; % file for topoopt==2
-smooth_topo=55; % n>0: smooth topography over n samples, if no smoothing =0
+% file should be in the same folder as radargrams.mat
+% topoopt=3: the topography is known at several points on the measured
+% area, but not directly on the profiles. The topofile contains 3 columns:
+% Easting, Northing and height
+topofile='Topo2d.txt'; % file for topoopt==2 or topoopt==3
+smooth_topo=55; % n>0: smooth topography over n samples, if no smoothing=0
 
 removeOutliers=0; % if =1: remove outliers in topo data (is done before smoothing)
 num=101; % odd(!) number of points for median calculation if removeOutliers==1
@@ -98,19 +100,14 @@ if vopt==1
     for i=1:length(global_coords)
         v{i}=vconst;
     end
-    maxdepth=max(t)/2*vconst; % maximum depth of radargram
 elseif vopt==2
     for i=1:length(global_coords)
         v{i}=vall(i);
     end
-    maxdepth=max(max(t)./2.*vall); % maximum depth of radargram
 elseif vopt==3
     for i=1:length(global_coords)
         v{i}=repmat(interp1(t1d,v1d,t,'linear',v1d(end)),[1 length(global_coords{i}(:,1))]);
-        % get depth for each profile
-        d(i)=max(t)/2*v{i}(end);
     end
-    maxdepth=max(d); % maximum depth of radargram
 end
 save(fullfile(pfad_rad,'vgrid.mat'),'v');
 
@@ -226,21 +223,54 @@ if topoopt>0
             global_coords{i}(:,3)=topo{i};
         end
         save(fullfile(pfad_rad,'global_coords.mat'),'global_coords');
+   elseif topoopt==3
+        top=load(fullfile(pfad_rad,topofile));
+        % interpolate topography from area to radargrams:
+        F=scatteredInterpolant(top(:,1),top(:,2),top(:,3),'linear','linear');
+        if smooth_topo>0
+            for i=1:length(global_coords)
+                topo{i}=F(global_coords{i}(:,1),global_coords{i}(:,2));
+                if plottopo==1
+                    figure
+                    plot(topo{i},'Marker','*')
+                    hold on
+                    xlabel('Trace number')
+                    ylabel('Topography [m]')
+                    led=[{'raw data'}];
+                    legend(led)
+                    grid on
[…]
