+++ GUIs/PondView_GUI.m
@@ -1,0 +1,825 @@
+function []=PondView_GUI(xgrid,ygrid,list,mask_interp,t,tz_flag,maxElevation,tr,radar,coords,pfad)
+
+% function [] = PondView_GUI(xgrid,ygrid,list,mask_interp,t,tz_flag,maxElevation,tr,radar,coords,pfad)
+%
+% Plot thickslices in PondView (Grasmuek & Viggiano 2018)
+%
+% Dr. Tina Wunderlich, CAU Kiel 2024, tina.wunderlich@ifg.uni-kiel.de
+%
+% Input:
+% xgrid, ygrid: Coordinate grids in m
+% list: list of slice-files
+% mask_interp: mask with 1 where data is
+% t: vector with of time/depth samples
+% tz_flag: =2 if depthslices instead of timeslices (tz_flag=1)
+% maxElevation: if tz_flag=2: Elevation (a.s.l) at depth=0m (maxElevation=[] for tz_flag=1)
+% tr: time/depth vector for radargrams
+% radar: radargrams in cells
+% coords: gloab_coords in cells for radargrams
+% pfad: path to slices folder
+%
+% requires folder Subfunctions
+
+S.pfad=pfad;
+
+% set data of slices
+S.x=xgrid; % in global coordinates
+S.y=ygrid;
+S.maxElevation=maxElevation;
+S.timedepth=tz_flag; % time/depth flag
+if S.timedepth==1
+    S.unit='ns';
+else
+    S.unit='m';
+end
+[S.r,S.c]=size(S.x);
+
+S.t = t; % time vector for slices
+S.dx=xgrid(1,2)-xgrid(1,1);
+S.dy=ygrid(2,1)-ygrid(1,1);
+S.mask=mask_interp;
+
+% initialize values for slices
+S.tt=0; % time/depth of starttime of slice (in ns or m)
+if S.timedepth==1
+    S.dz=10; % thickness of slice in ns
+else
+    S.dz=0.2; % thickness in m
+end
+S.thresh=0.2; % Amplitude threshold as min
+
+s=S.tt:S.dz/2:S.t(end)-S.dz/2;
+e=S.dz:S.dz/2:S.t(end);
+S.t_tsl(:,1)=s(1:min([length(s),length(e)])'); % start of thickslice
+S.t_tsl(:,2)=e(1:min([length(s),length(e)])'); % end of thickslice
+
+S.tslnum=1; % current slice number
+
+% load slices:
+disp('Loading slices...')
+S.temp=load(fullfile(list(1).folder,list(1).name));
+[r,c]=size(S.temp.slice); % size of slice 1
+S.n=length(list);
+S.slices=zeros(r,c,S.n);
+for i=1:S.n
+    temp=extractBetween(list(i).name,'_','.');
+    num=str2double(temp{1}); % number of slice from name
+    sl=load(fullfile(list(i).folder,list(i).name));
+    S.slices(:,:,num)=sl.slice; % set this slice at correct position
+end
+rmfield(S,'temp');
+
+
+
+% set radargramms (=data)
+S.data=radar;
+S.ar=1/10; % aspect ratio factor
+S.tr=tr; % time vector for radargrams
+if S.timedepth==2
+    S.tr=abs(S.tr-S.tr(1)); % depth is positive down, starting at 0m -> corresponds to dsl
+end
+
+for i=1:length(coords)
+    S.xprof{i}=coords{i}(:,1);  % x-coordinate along profile
+    S.yprof{i}=coords{i}(:,2);  % y-coordinate along profile
[…]
