+++ Visualization/PondViewer.m
@@ -1,0 +1,294 @@
+clear all
+close all
+clc
+
+% Use the binned 3D data blocks and visualize them with PondView (Grasmueck & Viggiano 2018)
+% M. Grasmueck and D. Viggiano, "PondView: Intuitive and Efficient Visualization of 3D GPR Data," 2018 17th International Conference on Ground Penetrating Radar (GPR), Rapperswil, Switzerland, 2018, pp. 1-6, doi: 10.1109/ICGPR.2018.8441634.
+%
+% Dr. Tina Wunderlich, CAU Kiel 2024, tina.wunderlich@ifg.uni-kiel.de
+%
+% requires PondView_GUI in folder GUIs and Subfunctions
+
+
+% time or depth domain?
+tz_flag=2; % =1: time, =2: depth
+
+rectangles=[1:2]; % number of rectangles
+
+radius=0.3; % radius in m for valid interpolation
+
+use_slices=1;   % if =0: make new slices and save them,
+% if =1: use already created slices in folder
+% PondViewSlices (but nevertheless select folder containing
+% the 3D_Grid... and PondViewSlices folders)
+
+%--------------------------------------------------------------------------
+% DO NOT CHANGE FROM HERE ON!
+warning off
+
+% path for slices/3D_Grids
+if ~ispc; menu('Choose folder containing 3D_Grid_R*-folders','OK'); end
+if ispc
+    if exist('temp.temp') % read last opened folder from temp.temp
+        fid=fopen('temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            pfad=uigetdir(fn{1}{1},'Select folder containing 3D_Grid_R*-folders');
+        else
+            pfad=uigetdir([],'Select folder containing 3D_Grid_R*-folders');
+        end
+        fileattrib('temp.temp','-h');
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',pfad);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    else
+        pfad=uigetdir([],'Select folder containing 3D_Grid_R*-folders');
+
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',pfad);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    end
+else
+    if exist('.temp.temp') % read last opened folder from temp.temp
+        fid=fopen('.temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            pfad=uigetdir(fn{1}{1},'Select folder containing 3D_Grid_R*-folders');
+        else
+            pfad=uigetdir([],'Select folder containing 3D_Grid_R*-folders');
+        end
+    else
+        pfad=uigetdir([],'Select folder containing 3D_Grid_R*-folders');
+    end
+
+    fid=fopen('.temp.temp','wt');
+    fprintf(fid,'%s',pfad);
+    fclose(fid);
+end
+
+% path for radargrams
+if ~ispc; menu('Choose folder with radargrams','OK'); end
+if ispc
+    if exist('rtemp.temp') % read last opened folder from temp.temp
+        fid=fopen('rtemp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            pfad_rad=uigetdir(fn{1}{1},'Select folder containing radargrams.mat');
+        else
+            pfad_rad=uigetdir([],'Select folder containing radargrams.mat');
+        end
+        fileattrib('rtemp.temp','-h');
[…]
