+++ Visualization/PondViewer.m
@@ -11,7 +11,7 @@ clc
 
 
 % time or depth domain?
-tz_flag=2; % =1: time, =2: depth
+tz_flag=1; % =1: time, =2: depth
 
 rectangles=[1:2]; % number of rectangles
 
