+++ Scripts_for_3D/Mala3D.m
@@ -19,10 +19,10 @@ clc
 platform=2; % Linux=1, Mac=2, Windows=3
 
 % Select number of profiles:
-profile_min=0;  % minimum profile number
-profile_max=2;  % maximum profile number
+profile_min=3;  % minimum profile number
+profile_max=6;  % maximum profile number
 % number of channels for this dataset
-channels=16; % number of channels
+channels=32; % number of channels
 
 changeDir=0; % if =1: change the sign of the y-antenna-GPS-offset, if =0: use offsets as written in file
 

[…]
