### 2024-04-26 b820bc7
General: small improvements regarding runtime. Additionally: (1) For long profiles/high data volume of Mala Mira data there is a new script for processing of the data including cross-channel t0 corrections (MalaProcessing.m), which is basically the first part of Mala3D. The output data only has single precision (and not double as before) to reduce data volume. (2) For binning of the processed data from MalaProcessing.m there is a new script (Bin_MalaProc.m), which is equal to the last part of Mala3D, but again only in single precision. (3) There is a script prepareRadargramList.m for creating 
[…]
