+++ Scripts_for_2D/MalaProcessing.m
@@ -3,7 +3,7 @@
 % If you measured with PPS/GPS-mouse for triggering, run convert_PPS2pos.m
 % first!
 %
-% Dr. Tina Wunderlich, CAU Kiel 2020, tina.wunderlich@ifg.uni-kiel.de
+% Dr. Tina Wunderlich, CAU Kiel 2024, tina.wunderlich@ifg.uni-kiel.de
 %
 % requires following files:
 % *.rd3: Radar data
@@ -22,7 +22,7 @@ clc
 platform=2; % Linux=1, Mac=2, Windows=3
 
 % Select number of profiles:
-profile_min=17;  % minimum profile number
+profile_min=0;  % minimum profile number
 profile_max=95;  % maximum profile number
 % number of channels for this dataset
 channels=16; % number of channels
@@ -46,8 +46,6 @@ userawdata=1;  % if =1: use aready read in raw data and apply new processing ste
 rad=0; % if =0, processed data will be saved in folder proc, but not all radargrams in one file radargrams.mat
 % if =1, processed data will be saved in radargrams.mat...
 
-% Save to sgy for Kingdom
-save_sgy=0; % yes=1, no=0
 
 %--------------------------------------------------------------------------
 % DO NOT CHANGE FROM HERE ON!
@@ -138,6 +136,7 @@ if ~exist(fullfile(foldername,settings),'file') % if no settings-file is found: 
     fprintf(fid,'do_applygain 0\ng1 -20\ng2 0\ng3 10\ng4 20\ng5 30\n\n');
     fprintf(fid,'do_normalization 0\nqclip 0.98\n\n');
     fprintf(fid,'do_removeMeanMedianTrace 7\nmeanmedian 1\nnumtraces 0\n\n');
+    fprintf(fid,'do_t0shift 0\nt0s[ns] 5.0\n\n');
     fprintf(fid,'do_t0CorrectionThreshold 0\nthreshold -1\n\n');
     fprintf(fid,'do_t0CorrectionReferencetrace 0\nprofilenumber 0\nchannelnumber 1\ntracenumber 1\nt0[ns] 0.1\n\n');
     fprintf(fid,'do_crossprofileshift 4\nmaxshift_cps 30\n\n');
@@ -151,7 +150,7 @@ else
     temp=textscan(fid,'%s%s');
     fclose(fid);
     % get order of processing steps:
-    steps=[{'do_amplitudeOffset'} {'do_crossprofileshift'} {'do_medianFilter'} {'do_constTraceDist'} {'do_cutTWT'} {'do_khighpass'} {'do_spectralWhitening'} {'do_migration2d_vz'} {'do_topomig2d'} {'do_traceInterpolation'} {'do_applygain'} {'do_removeMeanMedianTrace'} {'do_t0CorrectionThreshold'} {'do_t0CorrectionReferencetrace'} {'do_channelshift'} {'do_sphericalDivergence'} {'do_attenuationCorrection'} {'do_signalShaping'} {'do_bandpass'} {'do_normalization'}];
+    steps=[{'do_amplitudeOffset'} {'do_t0shift'} {'do_crossprofileshift'} {'do_medianFilter'} {'do_constTraceDist'} {'do_cutTWT'} {'do_khighpass'} {'do_spectralWhitening'} {'do_migration2d_vz'} {'do_topomig2d'} {'do_traceInterpolation'} {'do_applygain'} {'do_removeMeanMedianTrace'} {'do_t0CorrectionThreshold'} {'do_t0CorrectionReferencetrace'} {'do_channelshift'} {'do_sphericalDivergence'} {'do_attenuationCorrection'} {'do_signalShaping'} {'do_bandpass'} {'do_normalization'}];
[…]
