+++ Scripts_for_2D/prepareMigrationFiles.m
@@ -22,7 +22,7 @@ v1d=[0.16 0.1]; % v at different times in m/ns
 t1d=[0 18]; % corresponding times in ns
 
 % Choose options for topography:
-topoopt=1;
+topoopt=2;
 % topoopt=0: no topography file required
 % topoopt=1: topography is already set in global_coords(:,3)
 % topoopt=2: topography has to be set with file containing profile number (1. column),
@@ -31,14 +31,14 @@ topoopt=1;
 % topoopt=3: the topography is known at several points on the measured
 % area, but not directly on the profiles. The topofile contains 3 columns:
 % Easting, Northing and height
-topofile='Topo2d.txt'; % file for topoopt==2 or topoopt==3
+topofile='Topo200MHz.txt'; % file for topoopt==2 or topoopt==3
 smooth_topo=55; % n>0: smooth topography over n samples, if no smoothing=0
 
 removeOutliers=0; % if =1: remove outliers in topo data (is done before smoothing)
 num=101; % odd(!) number of points for median calculation if removeOutliers==1
 thresh=0.1; % threshold in m for difference between median and raw topography if removeOutliers==1
 
-plottopo=0; % if =1: topo is plotted (raw and smoothed), if =0: no figures
+plottopo=1; % if =1: topo is plotted (raw and smoothed), if =0: no figures
 
 %% -------------------------------------------------------------------------
 % Do not change the following part!
