+++ Visualization/saveFigures.m
@@ -21,7 +21,7 @@ aspectratio_t=10;    % for time plots: give aspectratio for y-axis. If you want 
 aspectratio_z=1;    % for depth plots: give aspectratio for y-axis. If you want to plot over whole screen, set =0. If you want to make the zaxis larger, make this number smaller.
 wiggleplot=0;  % =1: make wiggle plot instead of grayscale picture (=0)
 wigglescale=1; % scaling factor for wiggle plot
-tz_flag=2; % y-axis is 1=time [ns] or 2=depth [m]
+tz_flag=1; % y-axis is 1=time [ns] or 2=depth [m]
 ampSpec_flag=0; % 1: amplitude spectrum, 0: radargram
 dxtick=[]; % spacing between x-ticks in m (leave empty for automatic determination)
 dytick=[]; % spacing between y-ticks in ns or m (leave empty for automatic determination)
@@ -40,7 +40,7 @@ save_shape=1; % yes=1, no=0
 shapename='GPR_Profiles.shp'; % give name for shape file
 
 % Plotting options for map:
-plot_map=1; % plot map? yes=1, no=0
+plot_map=0; % plot map? yes=1, no=0
 map_xlim=[]; % give x-limits of map [start end] (leave empty for automatic determination)
 map_ylim=[]; % give y-limits of map [start end] (leave empty for automatic determination)
 xtick=[]; % spacing between x-ticks for map in m (leave empty for automatic determination)
