### 2024-05-03 56bdfce
added script extract_3Dblock.m for extracting a small block from the binned data
