+++ Scripts_for_3D/extract_3Dblock.m
@@ -1,0 +1,284 @@
+clear all
+close all
+clc
+
+
+% Use 3D-binned data and extract a smaller 3D-block from this. This block
+% will be saved in folder Small_3Dblock/3D_Grid_R1 (for compatibility with the other scripts also in a 3D_Grid_R1 folder).
+%
+% Dr. Tina Wunderlich, CAU Kiel 2024, tina.wunderlich@ifg.uni-kiel.de
+%
+% Requires binned 3D data in folders 3D_Grid_R*, if asked give rSlicer
+% folder (folder with 3D_Grid_R*-folders inside)
+
+
+rectangles=1:15; % give number of rectangles
+
+% give coordinate of lower left corner of small area, and width/height
+corner=[585254 5971678]; % Easting/Northing
+width=5; % width of small area (Easting/x) in m
+height=15; % height of small area (Northing/y) in m
+local_global=1; % =1: width/height in local(rotated) coordinate system; =2: width/height in global coordinate system
+
+proc=0;  % if you want to use processed 3D_blocks in 3D_Grid_R*/processed -> =1, else =0
+
+% Interpolation
+interp=1; % =1: do linear interpolation for every slice, =0: no interpolation
+radius=3; % radius for interpolation in bins
+
+% -------------------------------------------------------------------------
+% DO NOT CHANGE FROM HERE ON!
+warning('off');
+
+% get folder name 
+if ispc
+    if exist('temp.temp') % read last opened folder from temp.temp
+        fid=fopen('temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            pfad_rad=uigetdir(fn{1}{1},'Choose rSlicer-folder');
+        else
+            pfad_rad=uigetdir([],'Choose rSlicer-folder');
+        end
+        fileattrib('temp.temp','-h');
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',pfad_rad);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    else
+        pfad_rad=uigetdir([],'Choose rSlicer-folder'); % path to radargram-folder
+
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',pfad_rad);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    end
+else
+    if exist('.temp.temp') % read last opened folder from temp.temp
+        fid=fopen('.temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            pfad_rad=uigetdir(fn{1}{1},'Choose rSlicer-folder');
+        else
+            pfad_rad=uigetdir([],'Choose rSlicer-folder');
+        end
+    else
+        pfad_rad=uigetdir([],'Choose rSlicer-folder'); % path to radargram-folder
+    end
+
+    fid=fopen('.temp.temp','wt');
+    fprintf(fid,'%s',pfad_rad);
+    fclose(fid);
+end
+
+% set path temporarily:
+oldpath=path;
+currentFile = pwd;
+curFold=fileparts(currentFile);
+addpath(fullfile(curFold,'Subfunctions'));
+
+
+% read time vector
+if proc==0
[…]
