### 2024-05-08 e30a797
improved followTopography-option in make_Timeslices
