+++ Scripts_for_3D/make_Timeslices.m
@@ -11,25 +11,25 @@ clc
 % requires binned data in rectangles 3D_Grid_R* or processed data in
 % 3D_Grid_R*/processed/
 % Number of rectangles with binned data
-rectangles=1; % e.g. 1:3
+rectangles=1:22; % e.g. 1:3
 
 % depth slices instead of time slices (input is in m)
-dsl = 0; % =1: depth, =0: time
+dsl = 1; % =1: depth, =0: time
 
 % if depth slice, cut horizontally (=0) or follow topography (=1)?
-followTopo=0;
+followTopo=1;
 
 % starting time of first timeslice
 t_start=0;  % in ns (or m if depth slice starting from top of data (=0m))
 
 % thickness of timeslices
-thick=3; % in ns (or m if dsl=1)
+thick=0.2; % in ns (or m if dsl=1)
 
 % overlap of timeslices
 overlap = 0; % in ns (or m if dsl=1)
 
 % ending time of timeslices
-t_end=40; % in ns (or m if dsl=1, meters below top of data, positive!)
+t_end=3; % in ns (or m if dsl=1, meters below top of data, positive!)
 
 % dx of timeslices (<=dx of bins)
 dx_tsl=0.04;    % in m
@@ -165,6 +165,17 @@ for i=1:length(rectangles)
     z{i}=temp.z;
     temp=load(fullfile(pfad,p,'mask.mat'));
     mask{i}=temp.mask;
+    if isempty(mask{i})
+        if proc==0
+            p=['3D_Grid_R',int2str(rectangles(i)),'/data.mat'];
+        else
+            p=['3D_Grid_R',int2str(rectangles(i)),'/processed/data.mat'];
+        end
+        matFileObj=matfile(fullfile(pfad,p));
+        temp=matFileObj.data;
+        mask{i}=zeros(size(x{i}));
+        mask{i}(any(~isnan(temp),3))=1;
+    end
     
     if i==1
         dx=x{i}(1,2)-x{i}(1,1);   % dx of 3D-block
@@ -228,9 +239,9 @@ end
 tsl=cell(length(t_ind),1);  % initialize cells for timeslices
 
 disp('Reading data for timeslices... Please wait.');
-h=waitbar(0,'Reading data for timeslices');
 
 for j=1:length(rectangles)  % in each rectangle...
+    disp(['   Rectangle ',int2str(j),'/',int2str(length(rectangles))])
     % open matfile
     if proc==0
         p=['3D_Grid_R',int2str(rectangles(j)),'/data.mat'];
@@ -244,6 +255,7 @@ for j=1:length(rectangles)  % in each rectangle...
     col_ind=find(abs(xgrid(1,:)-xmin(j))==min(abs(xmin(j)-xgrid(1,:))));
     
     for i=1:length(t_ind)   % number of Tsl
+        disp(['      - Tsl ',int2str(i),'/',int2str(length(t_ind))])
         if isempty(tsl{i})
             tsl{i}=NaN(size(xgrid));    % initialize tsl in first run
         end
@@ -251,24 +263,42 @@ for j=1:length(rectangles)  % in each rectangle...
         if followTopo==0 % horizontal slices
             datatemp=matFileObj.data(:,:,t_ind{i});
         else % make Tsl following topography (for topocorr/mig data)
-            w=find(mask{j}==1); % linear indices of data points in rectangle
-            datatemp=NaN(length(z{j}(:,1)),length(z{j}(1,:)),length(t_ind{i}));
-            len_ind=zeros(size(z{j}))+length(t_ind{i});
-            for k=1:length(w) % go through all data points
-                [k1,k2]=ind2sub(size(z{j}),w(k));
[…]
