### 2024-05-16 c69cfc7
(1) added functions for reading multichannel sensors&software spidar data with processing and binning; (2) restructured the folder containing the scripts so that scripts for one specific equipment are in one folder. Also the scripts for Mala data formerly in Scripts_for_2D/3D are now in an extra folder.
