+++ SensorsSoftware_Import/Bin_SpidarProc.m
@@ -1,0 +1,558 @@
+% Script for reading processed Spidar-datafiles in profiles2mat/proc and
+% binning them onto a rectangular grid
+%
+% Dr. Tina Wunderlich, CAU Kiel 2024, tina.wunderlich@ifg.uni-kiel.de
+%
+% requires MATLAB-files in following folders (path will be temporarily
+% set):  Subfunctions
+%
+% Choose folder with original data files when asked.
+
+
+clear all
+close all
+clc
+
+% Computer system
+platform=2; % Linux=1, Mac=2, Windows=3
+
+% Bin size of grid
+dx=0.05; % [m]
+
+% Automatic rotation of measurement area for minimum memory size
+rotate_area=1;  % 1=yes, 0=no
+
+% Division into rectangles for processing
+force_division=0; % if =1: a division is forced although the computer memory is large enough for all data
+border=0.5;   % overlapping border of rectangles in m
+
+% Continue reading data for rectangles (if a former run of the program has
+% been stopped after some rectangles, you can continue here using the same
+% processed data and area configuration as before)
+continuedata=0; % if new area: =0
+continuerectangle=1;    % number of starting rectangle
+
+%--------------------------------------------------------------------------
+% DO NOT CHANGE FROM HERE ON!
+warning('off');
+
+% get folder name
+if ispc
+    if exist('temp.temp') % read last opened folder from temp.temp
+        fid=fopen('temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            foldername=uigetdir(fn{1}{1},'Choose folder with original data files');
+        else
+            foldername=uigetdir([],'Choose folder with original data files');
+        end
+        fileattrib('temp.temp','-h');
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',foldername);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    else
+        foldername=uigetdir([],'Choose folder with original data files'); % path to radargram-folder
+
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',foldername);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    end
+else
+    if exist('.temp.temp') % read last opened folder from temp.temp
+        fid=fopen('.temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            foldername=uigetdir(fn{1}{1},'Choose folder with original data files');
+        else
+            foldername=uigetdir([],'Choose folder with original data files');
+        end
+    else
+        foldername=uigetdir([],'Choose folder with original data files'); % path to radargram-folder
+    end
+
+    fid=fopen('.temp.temp','wt');
+    fprintf(fid,'%s',foldername);
+    fclose(fid);
+end
+
+
+% get name
+temp=dir(fullfile(foldername,'profiles2mat','*.mat'));
+name=extractBefore(temp(end).name,'_'); % Name of data files without ending
+
+
+% set path temporarily:
+oldpath=path;
+currentFile = pwd;
+curFold=fileparts(currentFile);
+addpath(fullfile(curFold,'Subfunctions'));
+
+% get matlab version (to shut down parfor_wait)
+rel=version('-release');
+if str2num(rel(1:4))<2017
+    pfw=0; % no parfor_wait
+else
+    pfw=1; % use parfor_wait
+end
+
+% get memory information of system
+if platform==1 % Linux
+    system('less /proc/meminfo >meminfo.txt');
[…]
