+++ SensorsSoftware_Import/Spidar_Processing.m
@@ -1,0 +1,1081 @@
+%------------------------ EKKO_Convert -------------------------------------
+%
+% Converts Spidar HD- & DT1- & GPS-files of Sensors&Software equipment to segy and mat (compatible with Multichannel-GPR)
+%
+% Dr. Tina Wunderlich, CAU Kiel, 2024, tina.wunderlich@ifg.uni-kiel.de
+%
+%--------------------------------------------------------------------------
+
+
+clear all
+close all
+clc
+
+name='NIC'; % name of data files
+
+% number of channels
+antenna_IDs=[1:4 6]; 
+gps_channel=1; % which channel is synchronized with the GPS? (-> same filename-number)
+
+offsetGPS_X=[-0.5,-0.25,0,0.25,0.5]; % Offset between GPS and antenna midpoint crossline (for each antenna midpoint, same order as antenna_IDs) (in profile direction GPS left of antenna -> positive)
+offsetGPS_Y=[0,0,0,0,0]; % Offset between GPS and antenna midpoint in profile direction (for each antenna midpoint, same order as antenna_IDs) (if GPS behind antenna midpoint -> positive)
+zone=32; % UTM zone (for conversion to WGS84/UTM)
+
+dataplot=0; % plot radargram for controlling? 1=yes, 0=no (Use with caution as every profile and every channel is plotted individually!)
+
+settings='settings.txt'; % name of file for processing settings. 
+% If this file does not exist, it will be created. You can modify it with a
+% normal text editor and run the script again.
+
+% Raw data is only read if userawdata=0! Otherwise mat-files of raw data in folder profiles2mat are used.
+userawdata=0;  % if =1: use aready read in raw data and apply new processing steps
+
+% Export to other formats
+export2mat=1; % export to Multichannel-GPR format for radargrams (mat-files) 
+% (do not use for large datasets here. Instead you can export it later using profiles2radargrams e.g. for single channels)
+
+%---------------------------- DO NOT CHANGE FROM HERE ON ----------------------------
+warning('off');
+
+% set path temporarily:
+oldpath=path;
+currentFile = pwd;
+curFold=fileparts(currentFile);
+addpath(fullfile(curFold,'Export_Import'),fullfile(curFold,'Subfunctions'),fullfile(curFold,'Processing'));
+
+
+% get file names
+if ispc
+    if exist('spidar.temp') % read last opened folder from temp.temp
+        fid=fopen('spidar.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            pfad=uigetdir(fn{1}{1},'Choose folder with spidar-file(s)');
+        else
+            pfad=uigetdir([],'Choose folder with spidar-file(s)');
+        end
+        fileattrib('spidar.temp','-h');
+        fid=fopen('spidar.temp','wt');
+        fprintf(fid,'%s',pfad);
+        fclose(fid);
+        fileattrib('spidar.temp','+h');
+    else
+        pfad=uigetdir([],'Choose folder with spidar-file(s)'); % path to radargram-folder
+
+        fid=fopen('spidar.temp','wt');
+        fprintf(fid,'%s',pfad);
+        fclose(fid);
+        fileattrib('spidar.temp','+h');
+    end
+else
+    if exist('.spidar.temp') % read last opened folder from temp.temp
[…]
