### 2024-05-16 #6248 @martinfleis
Hi @mrava87, would you be willing to review this submission for JOSS?
