### 2024-05-16 89dc7be
removed temporary working script
