+++ SensorsSoftware_Import/Spidar_Convert.m
@@ -1,351 +1,0 @@
-%------------------------ EKKO_Convert -------------------------------------
-%
-% Converts Spidar HD- & DT1- & GPS-files of Sensors&Software equipment to segy and mat (compatible with Multichannel-GPR)
-%
-% Dr. Tina Wunderlich, CAU Kiel, 2024, tina.wunderlich@ifg.uni-kiel.de
-%
-%--------------------------------------------------------------------------
-% STILL UNDER CONSTRUCTION! PLEASE DO NOT USE!
-
-clear all
-close all
-clc
-
-name='NIC'; % name of data files
-
-% number of channels
-antenna_IDs=[1:4 6]; 
-gps_channel=1; % which channel is synchronized with the GPS? (-> same filename-number)
-
-dataplot=1; % plot radargram for controlling? 1=yes, 0=no
-
-offsetGPS_X=[-0.5,-0.25,0,0.25,0.5]; % Offset between GPS and antenna midpoint crossline (for each antenna midpoint, same order as antenna_IDs) (in profile direction GPS left of antenna -> positive)
-offsetGPS_Y=[0,0,0,0,0]; % Offset between GPS and antenna midpoint in profile direction (for each antenna midpoint, same order as antenna_IDs) (if GPS behind antenna midpoint -> positive)
-
-% Options for calculating inline coordinates for each trace:
-coords_opt=3;   % =1: trace coordinate is difference to beginning of profile (only use this for straight profiles!)
-                % =2: trace coordinates are calculated by taking the cumulative sum of the coordinate differences between subsequent traces (better for curvy profiles, but not useful for strong GPS-antenna movements)
-                % =3: use distance in GPS-file
-
-% Export to other formats
-export2mat=1; % export to Multichannel-GPR format for radargrams (mat-files)
-export2segy=0; % export all radargrams as segy-files
-constoff=0; % for sgy: if=1: a constant coordinate offset will be subtracted and coordinates will be in mm accuracy in segy file (offsets will be saved in Inline3D (x) and Crossline3D (y))
-
-
-%---------------------------- DO NOT CHANGE FROM HERE ON ----------------------------
-% 
-% LOAD DATA
-
-% set path temporarily:
-oldpath=path;
-currentFile = pwd;
-curFold=fileparts(currentFile);
-addpath(fullfile(curFold,'Export_Import'),fullfile(curFold,'Subfunctions'));
-
-
-% get file names
-if ispc
-    if exist('spidar.temp') % read last opened folder from temp.temp
-        fid=fopen('spidar.temp','r');
-        fn=textscan(fid,'%s');
-        fclose(fid);
-        if ~isempty(fn{1})
-            pfad=uigetdir(fn{1}{1},'Choose folder with spidar-file(s)');
-        else
-            pfad=uigetdir([],'Choose folder with spidar-file(s)');
-        end
-        fileattrib('spidar.temp','-h');
-        fid=fopen('spidar.temp','wt');
-        fprintf(fid,'%s',pfad);
-        fclose(fid);
-        fileattrib('spidar.temp','+h');
-    else
-        pfad=uigetdir([],'Choose folder with spidar-file(s)'); % path to radargram-folder
-
-        fid=fopen('spidar.temp','wt');
-        fprintf(fid,'%s',pfad);
-        fclose(fid);
-        fileattrib('spidar.temp','+h');
-    end
-else
[…]
