### 2024-05-21 #6767 @martinfleis
All reviewers should create checklists with the JOSS requirements using the command `@editorialbot generate my checklist`. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues (and small pull requests if needed) on the software repository. When doing so, please mention `https://github.com/openjournals/joss-reviews/issues/6767 ` so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 2-4 weeks, feel free to start whenever it works for you. Please let me know if any of you require significantly more time at any point. We can also use `editorialbot` to set automatic reminders if you know you'll be away for a known period of time.

Thanks!
