### 2024-05-23 b24427d
(1) profiles2radargrams is now also working for spidar data; (2) Process_SingleProfile is now also working for spidar data; (3) Spidar_Processing is now getting automatically the file name and gps channel of the data and has the same processing options as in Process_SingleProfile (setting.txt are compatible).
