+++ Scripts_for_3D/make_Timeslices.m
@@ -29,7 +29,7 @@ thick=5; % in ns (or m if dsl=1)
 overlap = 0; % in ns (or m if dsl=1)
 
 % ending time of timeslices
-t_end=80; % in ns (or m if dsl=1, meters below top of data, positive!)
+t_end=60; % in ns (or m if dsl=1, meters below top of data, positive!)
 
 % dx of timeslices (<=dx of bins)
 dx_tsl=0.05;    % in m
