### 2024-05-29 abf8d43
corrected bug in make_Timeslices for followTopo-option
