+++ Visualization/saveFigures.m
@@ -13,15 +13,15 @@ clc
 % Dr. Tina Wunderlich, CAU Kiel 2021, tina.wunderlich@ifg.uni-kiel.de
 
 
-numbers=[]; % give numbers of processed radargrams or leave empty =[] for all
+numbers=[1]; % give numbers of processed radargrams or leave empty =[] for all
 
 % Plotting options for radargrams:
 colorclip=3; % 0 is colorscale from min(data) to max(data), 1 is 1% clip value, 2 is 2% clip value and 3 is 3% clip value, ... (will not be saved, for plotting only!)
-aspectratio_t=10;    % for time plots: give aspectratio for y-axis. If you want to plot over whole screen, set =0. If you want to make the taxis larger, make this number smaller.
+aspectratio_t=7;    % for time plots: give aspectratio for y-axis. If you want to plot over whole screen, set =0. If you want to make the taxis larger, make this number smaller.
 aspectratio_z=1;    % for depth plots: give aspectratio for y-axis. If you want to plot over whole screen, set =0. If you want to make the zaxis larger, make this number smaller.
 wiggleplot=0;  % =1: make wiggle plot instead of grayscale picture (=0)
 wigglescale=1; % scaling factor for wiggle plot
-tz_flag=2; % y-axis is 1=time [ns] or 2=depth [m]
+tz_flag=1; % y-axis is 1=time [ns] or 2=depth [m]
 ampSpec_flag=0; % 1: amplitude spectrum, 0: radargram
 dxtick=[]; % spacing between x-ticks in m (leave empty for automatic determination)
 dytick=[]; % spacing between y-ticks in ns or m (leave empty for automatic determination)
@@ -37,7 +37,7 @@ save_georef=1; % yes=1, no=0
 
 % save profile coordinates as shape file?
 save_shape=1; % yes=1, no=0
-shapename='GPR_Profiles_Chan8.shp'; % give name for shape file
+shapename='GPR_Profiles_Chan1.shp'; % give name for shape file
 
 % Plotting options for map:
 plot_map=0; % plot map? yes=1, no=0
