### 2024-05-29 bfc7eba
corrected small change in getting markers for new SIR4000 software update
