### 2024-05-31 140954e
corrected bug in readspidar for handling some inconsistencies in the GPS-files
