### 2024-06-12 c358c0c
corrected readdzt_4000 for different SIR4000 software versions and marker/GPS measurements
