+++ GSSI_Import/DZT_Convert.m
@@ -15,7 +15,7 @@ app='SIR4000';    % Equipment: SIR20 / SIR30 / SIR3000 / SIR4000 / Tablet
 
 dataplot=1; % plot radargram for controlling? 1=yes, 0=no
 
-convert2utm=0; % convert WGS Lat/Long to UTM (=1 if measured with Stonex-GPS)
+convert2utm=1; % convert WGS Lat/Long to UTM (=1 if measured with Stonex-GPS)
 zone=32; % if convert2utm==1 -> give UTM-zone
 
 offsetGPS_X=0; % [m] Offset between GPS and antenna midpoint crossline (in profile direction GPS left of antenna -> positive)
