### 2024-06-13 7032ad3
added output of GPS quality flag when using SIR4000 with GPS; some general bug fixes for GSSI import for SIR20 and SIR3000
