+++ GSSI_Import/DZT_Convert.m
@@ -15,8 +15,8 @@ app='SIR4000';    % Equipment: SIR20 / SIR30 / SIR3000 / SIR4000 / Tablet
 
 dataplot=1; % plot radargram for controlling? 1=yes, 0=no
 
-convert2utm=1; % convert WGS Lat/Long to UTM (=1 if measured with Stonex-GPS)
-zone=32; % if convert2utm==1 -> give UTM-zone
+convert2utm=0; % convert WGS Lat/Long to UTM (=1 if measured with Stonex-GPS)
+zone=33; % if convert2utm==1 -> give UTM-zone
 
 offsetGPS_X=0; % [m] Offset between GPS and antenna midpoint crossline (in profile direction GPS left of antenna -> positive)
 offsetGPS_Y=0; % [m] Offset between GPS and antenna midpoint in profile direction (if GPS behind antenna midpoint -> positive)
@@ -138,7 +138,7 @@ for i=1:length(list)
         elseif strcmp(app,'SIR30')
             [data,trh,h]=readdzt_30(fullfile(pfad,[name{i},'.DZT']),dataplot);
         else
-            [data,trh,h]=readdzt_all(fullfile(pfad,[name{i},'.DZT']),dataplot);
+            [data,trh,h]=readdzt_all(fullfile(pfad,[name{i},'.DZT']),dataplot,app);
         end
         
         if ~strcmp(app,'SIR4000') && convert2utm==1 && exist([pfad,'/', name{i}, '.DZG'])
@@ -165,6 +165,7 @@ for i=1:length(list)
             trh.y=trh.y(~weg);
             trh.z=trh.z(~weg);
             trh.time=trh.time(~weg);
+            trh.quality=trh.quality(~weg);
             trh.mark=trh.mark(~weg);
             trh.channum=trh.channum(~weg);
             trh.tracenum=1:length(trh.mark);
@@ -416,6 +417,7 @@ for i=1:length(list)
         % if readdzt_all was used no field time is created
         if ~isfield(trh,'time')
             trh.time = zeros(size(trh.x));
+            trh.quality = NaN(size(trh.x)); % NaN because zero has a definition: (0=Fix not valid)
         end
         
         if h.nchan==2
@@ -438,8 +440,10 @@ for i=1:length(list)
                 elseif coords_opt==2 % cumulative sum
                     x=[x; {[0; cumsum(sqrt(diff(xtemp).^2+diff(ytemp).^2)')]}];
                 end
-                time = [{trh.time(trh.channum==1) }];
-                time = [time; {trh.time(trh.channum==2) }];
+                time = [{trh.time(trh.channum==1)}];
+                time = [time; {trh.time(trh.channum==2)}];
+                quality = [{trh.quality(trh.channum==1)}];
+                quality = [quality; {trh.quality(trh.channum==2)}];
                 if ~all(trh.mark==0) % marker present
                     marker=[{trh.mark(trh.channum==1)}];
                     marker=[marker; {trh.mark(trh.channum==2)}];
@@ -465,8 +469,10 @@ for i=1:length(list)
                 elseif coords_opt==2 % cumulative sum
                     x=[x; {[0; cumsum(sqrt(diff(xtemp).^2+diff(ytemp).^2)')]}];  
                 end
-                time = [time; {trh.time(trh.channum==1) }];
-                time = [time; {trh.time(trh.channum==2) }];
+                time = [time; {trh.time(trh.channum==1)}];
+                time = [time; {trh.time(trh.channum==2)}];
+                quality = [quality; {trh.quality(trh.channum==1)}];
+                quality = [quality; {trh.quality(trh.channum==2)}];
                 if ~all(trh.mark==0) % marker present
[…]
