### 2024-06-13 d5933d9
added script MalaProcessing_withoutGPS.m which can import and process data if only start and end point of the profile are known (and if there is no rSlicer folder)
