+++ Export_Import/readmala_withoutGPS.m
@@ -1,0 +1,219 @@
+function [traces,dt,ns,x,y,z,numchannels]=readmala_withoutGPS(foldername,name,profile_num)
+
+% Read Mala rd7-data (each channel in one file)
+% [traces,dt,ns,x,y,z,numchannels]=readmala_withoutGPS(foldername,name,profile_num)
+%
+% Dr. Tina Wunderlich, CAU Kiel 2024, tina.wunderlich@ifg.uni-kiel.de
+%
+% Input: 
+% foldername: complete rSlicer-foldername and path
+% name: Name of datafiles without '_...'
+% profile_num: Numbers of profiles
+% changeDir: flag if you want to change the +/- of the
+% y-antenna-GPS-offset, yes=1, no=0
+% add_Yoffset: additional offset for y-antenna-GPS-offset
+%
+% Output:
+% traces is matrix of size [ns,numchannels*numtraces_per_channel] 
+% (order is: first all traces of 1. channels, all traces of 2. channel,....)
+% x, y, z: vectors of length [1,numchannels*numtraces_per_channel] with
+% coordinates for each trace
+% dt: sampling interval in ns
+% ns: number of samples per trace
+% numchannels: number of channels
+%
+% requires helmert.m
+
+
+%%% Read yml-file
+if profile_num<10
+    temp=readlines(fullfile(foldername,[name,'_00',int2str(profile_num),'.yml']));
+elseif profile_num>=10 && profile_num<100
+    temp=readlines(fullfile(foldername,[name,'_0',int2str(profile_num),'.yml']));
+else
+    temp=readlines(fullfile(foldername,[name,'_',int2str(profile_num),'.yml']));
+end
+% get channel infos:
+ind=find(strcmp(temp,'SELECTED CHANNELS:')); % find this line
+anz=1;
+while ~strcmp(temp(ind+anz),'TRANSMITTERS:')
+    ch{anz}=extractAfter(temp(ind+anz),'- ');
+    anz=anz+1;
+end
+test=cellfun(@(x) strsplit(x,':'),ch,'UniformOutput',false); % split R* and T*
+channelinfo=zeros(length(test),3); % channel# T# R#
+channelinfo(:,1)=1:length(test); % channel number
+channelinfo(:,2)=cellfun(@(x) str2num(extractAfter(x{1},'T')),test); % transmitter #
+channelinfo(:,3)=cellfun(@(x) str2num(extractAfter(x{2},'R')),test); % receiver #
+
+% get positions of transmitters:
+ind=find(strcmp(temp,'TRANSMITTERS:')); % find this line
+i=1;
+while ~strcmp(temp(ind+(i-1)*3+1),'RECIVERS:')
+    Tx(i,1)=str2num(extractAfter(temp(ind+(i-1)*3+2),'X-POS: '));
+    Ty(i,1)=str2num(extractAfter(temp(ind+(i-1)*3+3),'Y-POS: '));
+    i=i+1;
+end
+
+% get positions of receivers:
+ind=find(strcmp(temp,'RECIVERS:')); % find this line
+i=1;
+while ~strcmp(temp(ind+(i-1)*3+1),'WHEEL:')
+    Rx(i,1)=str2num(extractAfter(temp(ind+(i-1)*3+2),'X-POS: '));
+    Ry(i,1)=str2num(extractAfter(temp(ind+(i-1)*3+3),'Y-POS: '));
+    i=i+1;
+end
+
+
+
+%%% Read header file
+if profile_num<10
+    list=dir(fullfile(foldername,[name,'_00',int2str(profile_num),'_*.rad'])); % list of all rad-files for this profile (one for each channel)
[…]
