+++ Mala_Import/MalaProcessing_withoutGPS.m
@@ -1,0 +1,1046 @@
+% Script for reading Mala-datafiles (NOT in rSlicer folder but all channels individually)
+% and processing them (no binning!)
+%
+% Dr. Tina Wunderlich, CAU Kiel 2024, tina.wunderlich@ifg.uni-kiel.de
+%
+% requires following files:
+% *.rd7: Radar data of individual channels
+% *.rad: header file
+% *.pos: positioning file with position for start and end point of the
+% profile (same format as the "normal" pos-files in the rSlicer-folder, but
+% just with one coordinate for first and last trace).
+%
+% requires MATLAB-files in following folders (path will be temporarily
+% set): Processing, Export_Import, Subfunctions, Migration
+
+
+clear all
+close all
+clc
+
+% Computer system
+platform=2; % Linux=1, Mac=2, Windows=3
+
+% Select number of profiles:
+profile_min=1;  % minimum profile number
+profile_max=17;  % maximum profile number
+% number of channels for this dataset
+channels=22; % number of channels
+
+changeDir=0; % if =1: change the sign of the y-antenna-GPS-offset, if =0: use offsets as written in file
+
+add_Yoffset=0;    % add a constant offset to the y-antenna-GPS-offset (e.g. due to non-vertical GPS-stick), set =0 if not applicable
+% negative if GPS is tilted towards front, positive if
+% tilted towards back
+
+%%% Processing steps before binning in txt-file (a default file with this
+%%% name is created, if not existing)
+settings='settings.txt';
+
+%%% Raw data is only read if userawdata=0! Otherwise mat-files of raw
+%%% data in folder profiles2mat are used.
+userawdata=1;  % if =1: use aready read in raw data and apply new processing steps
+
+% save radargrams.mat of processed profile data in one variable (requires large memory -> probably not
+% working for every data set)?
+rad=0; % if =0, processed data will be saved in folder proc, but not all radargrams in one file radargrams.mat
+% if =1, processed data will be saved in radargrams.mat...
+
+
+%--------------------------------------------------------------------------
+% DO NOT CHANGE FROM HERE ON!
+warning('off');
+
+% get folder name
+if ispc
+    if exist('temp.temp') % read last opened folder from temp.temp
+        fid=fopen('temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            foldername=uigetdir(fn{1}{1},'Choose folder with data');
+        else
+            foldername=uigetdir([],'Choose folder with data');
+        end
+        fileattrib('temp.temp','-h');
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',foldername);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    else
+        foldername=uigetdir([],'Choose folder with data'); % path to radargram-folder
+
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',foldername);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    end
+else
+    if exist('.temp.temp') % read last opened folder from temp.temp
+        fid=fopen('.temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
[…]

[…]
