### 2024-07-08 3f79a62
Topomigration: if there are missing (=NaN) traces, the radargram will be split into parts and each part will be migrated individually and put together afterwards
