+++ Migration/topomig2d_varV.m
@@ -74,16 +74,16 @@ if flag==1
     if verbose==1
         disp('Topographic correction')
     end
-    
+
     % initialize data matrix
     dataz=NaN(nsz,nt);
-    
+
     % for constant v:
     if all(size(v)==1)
         for j=1:nt % go through all traces
             dataz(round((gzmax-topo(j))/dz):round((gzmax-topo(j))/dz)+ns-1,j)=data(:,j);
         end
-    % for variable v:
+        % for variable v:
     else
         for j=1:nt % go through all traces
             ztemp=t./2.*v(:,j); % use v at this trace for depth conversion
@@ -92,133 +92,153 @@ if flag==1
             dataz(round((gzmax-topo(j))/dz):round((gzmax-topo(j))/dz)+length(neu)-1,j)=interp1(ztemp,data(:,j),neu);
         end
     end
-    
+
 else % Topomigration
     if verbose==1
         disp('Topo migration')
     end
-    
+
     if length(topo(:,1))<length(topo(1,:))
         topo=topo'; % make topo a column vector
     end
-    
+
     if all(size(v)==1) % for constant v -> make vgrid
         v=zeros(size(data))+v;
     end
-    
+
     % Estimate width of aperture to reduce matrix size later
     maxi_z=t(end)./2.*max(v(:)); % max. penetration depth [m]
     l=2*pi*maxi_z/360*aperture; % length of semicircle for aperture width and maximum depth [m]
     width=ceil(max(maxi_z,l)).*2; % width of area for matrix reduction [m]
-    
-    % pad radargram (and corresponding matrices) with zeros in the beginning and end to reduce edge
-    % effects
-    data=[zeros(length(t),round((width/2/dx))) data zeros(length(t),round((width/2/dx)))];
-    v=[zeros(length(t),round((width/2/dx))) v zeros(length(t),round((width/2/dx)))];
-    x=[zeros(1,round((width/2/dx))) x zeros(1,round((width/2/dx)))];
-    topo=[zeros(1,round((width/2/dx))) topo' zeros(1,round((width/2/dx)))];
-    los=find(data(10,:)~=0,1,'first'); % index for starting of original radargram...
-    bis=find(data(10,:)~=0,1,'last'); % ...index for ending of original radargram
-    
-    nt2=length(x); % number of traces for padded radargram
-    
-    % initialize big matrix
-    mat=zeros(nsz,nt2); % matrix for amplitude summation
-    count=zeros(nsz,nt2); % counter matrix
-    
-    for j=los:bis-1 % go through traces (without padded area)
-        if verbose==1
-            if mod(j-los+1,100)==0
-                fprintf('Trace: %d\n',j-los+1)
-            end
+
+    % if there are missing (NaN) traces inbetween -> split radargram here
+    ind=find(isnan(data(1,:))); % find NaN-Traces
+    diffind=diff(ind); % find tracenumber-differences between NaN-traces
+    if ~isempty(ind) % there are nan-traces
+        di=find(diffind~=1);    % jumps
+        part=[1 ind(1)-1]; % start-end of part with data
+        for i=1:length(di)
+            part=[part; ind(di(i))+1 ind(di(i)+1)-1];
         end
-        
-        % current gradient of topo
-        dtopo=topo(j+1)-topo(j);
-        % Angle of radiation (0° = vertically down)
-        if dtopo<0
-            alpha=atand(dtopo/dx);
-        elseif dtopo==0
-            alpha=0;
-        else
-            alpha=atand(dtopo/dx);
-        end
-        
-        % calculate radius for each Sample on this trace (depending on v
-        % in this trace)
-        rad=t./2.*v(:,j); % in m
-        
-        ind_z=find(abs(z-topo(j))==min(abs(z-topo(j))),1,'first'); % get z-index of surface point in this trace
[…]
