### 2024-07-10 43aabe8
added zoom function for layerpicking
