+++ GUIs/plot_layerpicking.m
@@ -61,8 +61,10 @@ S.fh.Visible='on';
         S.panel1a=uipanel('Parent',S.panel1);
         set(S.panel1,'Position',[210 80+S.siz(4)/2+80 S.siz(3)-220 S.siz(4)/2-100],'Units','pixels'); % panel for cutting frame with slider
         set(S.panel1a,'Position',[210 80+S.siz(4)/2+80 2000 S.siz(4)/2-100],'Units','pixels'); % panel inside cutting frame with image
+        S.panel1a_pos=[210 80+S.siz(4)/2+80 2000 S.siz(4)/2-100];
 
-        S.slider1=uicontrol('Style','Slider','Parent',S.fh,'Units','pixels','Position',[210 80+S.siz(4)/2+80 S.siz(3)-220 0.05],'Value',0.5,'Callback',@slider_callback1);
+        S.slider1H=uicontrol('Style','Slider','Parent',S.fh,'Units','pixels','Position',[210 80+S.siz(4)/2+80 S.siz(3)-220 0.05],'Value',0.5,'Callback',@slider_callback1);
+        S.slider1V=uicontrol('Style','Slider','Parent',S.fh,'Units','pixels','Position',[210 120 0.05 S.siz(4)/2-100],'Value',0.5,'Callback',@slider_callback1);
         
         S.ax1 = axes('OuterPosition',[0.02 0.02 0.98 0.98],'Units','normalized','Parent',S.panel1a);
         % make initial plot
@@ -94,9 +96,12 @@ S.fh.Visible='on';
         S.panel2a=uipanel('Parent',S.panel2);
         set(S.panel2,'Position',[210 120 S.siz(3)-220 S.siz(4)/2-100],'Units','pixels');
         set(S.panel2a,'Position',[210 120 2000 S.siz(4)/2-100],'Units','pixels');
+        S.panel2a_pos=[210 120 2000 S.siz(4)/2-100];
 
-        S.slider2=uicontrol('Style','Slider','Parent',S.fh,'Units','pixels','Position',[210 120 S.siz(3)-220 0.05],'Value',0.5,'Callback',@slider_callback2);
-        
+        S.slider2H=uicontrol('Style','Slider','Parent',S.fh,'Units','pixels','Position',[210 120 S.siz(3)-220 0.05],'Value',0.5,'Callback',@slider_callback2);
+        S.slider2V=uicontrol('Style','Slider','Parent',S.fh,'Units','pixels','Position',[210 80+S.siz(4)/2+80 0.05 S.siz(4)/2-100],'Value',0.5,'Callback',@slider_callback2);
+
+
         S.ax2 = axes('OuterPosition',[0.02 0.02 0.98 0.98],'Units','normalized','Parent',S.panel2a);
         % make initial plot
         if isnan(S.z)
@@ -148,22 +153,25 @@ S.fh.Visible='on';
         S.flag=1;
         
         % UI control - pushbutton for saving
-        S.save=uicontrol(S.fh,'Style','pushbutton','String','Save all picks','Position',[10 220 180 25],'Callback',@save_picks);
+        S.save=uicontrol(S.fh,'Style','pushbutton','String','Save all picks','Position',[10 320 180 25],'Callback',@save_picks);
         
         % UI control - pushbutton for loading
-        S.load=uicontrol(S.fh,'Style','pushbutton','String','Load picks','Position',[10 250 180 25],'Callback',@load_picks);
+        S.load=uicontrol(S.fh,'Style','pushbutton','String','Load picks','Position',[10 350 180 25],'Callback',@load_picks);
         
         % UI control - pushbutton for deleting of layer
-        S.dellayer=uicontrol(S.fh,'Style','pushbutton','String','Delete current layer','Position',[10 280 180 25],'Callback',@delete_layer);
[…]
