### 2024-07-15 32b3e8d
corrected bug in convert_PPS2pos
