### 2024-07-17 2bd2fb7
converted time to ns for DF antenna in sgy export
