+++ GSSI_Import/DZT_Convert.m
@@ -35,7 +35,7 @@ removeOutliers=0; % do you want to remove coordinate outliers?
 
 % Export to other formats
 export2mat=1; % export to Multichannel-GPR format for radargrams (mat-files)
-export2segy=0; % export all radargrams as segy-files
+export2segy=1; % export all radargrams as segy-files
 constoff=0; % if=1: a constant coordinate offset will be subtracted and coordinates will be in mm accuracy in segy file (offsets will be saved in Inline3D (x) and Crossline3D (y))
 
 
@@ -832,6 +832,11 @@ if export2segy==1
     % make new folder
     if ~exist(fullfile(pfad,'SEGY'),'dir')
         mkdir(fullfile(pfad,'SEGY'));
+    end
+
+    if isfield(h,'dt2') % is DF antenna
+        h.dt=h.dt*1e9; % convert to ns
+        h.dt2=h.dt2*1e9;
     end
     
     for i=1:length(radargrams)
