### 2024-07-24 71d2efd
added kk-pieslicefilter for reducing stripe noise in timeslices
