+++ Scripts_for_3D/Timeslices_kk_pieslicefilter.m
@@ -1,0 +1,280 @@
+clear all
+close all
+clc
+
+%% kk-Pieslicefilter for Timeslices/Depthslices with selection of different areas (using a shape-file)
+%
+% Dr. Tina Wunderlich, CAU Kiel 2024, tina.wunderlich@ifg.uni-kiel.de
+%
+% requires interpolated Timeslices in MultichannelGPR-format (created with
+% make_Timeslices.m)
+% (requires scripts in "Subfunctions")
+%
+% When asked, choose folder with interpolated timeslices (containing tsl_interp.mat, xgrid_interp.mat, ...)
+
+
+% name of shapefile for different parts of the area (in same folder as the timeslices):
+shpfile='kkFilterShape.shp'; % needs to be a shape-file with polygons of the different areas for filtering 
+all_area=0; % if =1 the complete area will be used with one stripe noise direction and shpfile will be ignored (for using shpfile set to 0)
+
+% give name for new folder with filtered results and settings (will be
+% created inside the interpolated timeslices folder)
+name='kkfilt';
+
+use_old_settings=1; % if =1: use old settings (angles of stripe noise from folder "name"), =0: make new angles
+
+apply_to_all_tsl=0; % =1: if you want to appyl it to all timeslices and save in MultichannelGPR-format
+
+apply_to_tslnum=1; % =1: if you want to apply it to tslnum only and save comparison figure
+tslnum=14; % if you want to make new angles (use_old_settings=0), choose number of timeslice for plotting and interactive picking of stripe noise
+
+opening=20; % opening angle of pieslice filter in degree
+
+% plotting options:
+colorclip=3;
+
+plotflag=1; % do you want to plot original spectrum and filtered  spectrum (only for tslnum)? yes=1, no=0
+
+
+%% ------------- DO NOT CHANGE FROM HERE ON! ----------------------------- %%
+warning('off');
+
+% get folder name
+if ispc
+    if exist('temp.temp','file') % read last opened folder from temp.temp
+        fid=fopen('temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            folder=uigetdir(fn{1}{1},'Choose interpolated timeslices folder');
+        else
+            folder=uigetdir([],'Choose interpolated timeslices folder');
+        end
+        fileattrib('temp.temp','-h');
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',folder);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    else
+        folder=uigetdir([],'Choose interpolated timeslices folder'); % path to timeslices-folder
+
+        fid=fopen('temp.temp','wt');
+        fprintf(fid,'%s',folder);
+        fclose(fid);
+        fileattrib('temp.temp','+h');
+    end
+else
+    if exist('.temp.temp','file') % read last opened folder from temp.temp
+        fid=fopen('.temp.temp','r');
+        fn=textscan(fid,'%s');
+        fclose(fid);
+        if ~isempty(fn{1})
+            folder=uigetdir(fn{1}{1},'Choose interpolated timeslices folder');
+        else
+            folder=uigetdir([],'Choose interpolated timeslices folder');
+        end
+    else
[…]
