### 2024-07-24 16a35b1
added kk-pieslicefilter for reducing stripe noise in timeslices - corrected small bug
