### 2024-08-06 #6767 @lrverdon
Hi @martinfleis,

I have nearly completed the review (sorry for the late submission, due to a long field campaign abroad)

I have a few minor problems with the software, which I could submit as an issue in the software repository, and also a list of small comments (e.g. typos, small errors or things that could be clarified in the software paper or in the Wiki in the repository). How is it best to communicate these? Should I add them below as comments in this review?
Thank you.
