### 2024-09-18 41e4ad4
added a GUI for CMP data processing: folder CMP_Processing/CMP_Processing. YOu can manually fit hyperbola or calculate semblance
