+++ GUIs/CMPproc_exported.m
@@ -1,0 +1,940 @@
+classdef CMPproc_exported < matlab.apps.AppBase
+
+    % Properties that correspond to app components
+    properties (Access = public)
+        UIFigure                 matlab.ui.Figure
+        GridLayout               matlab.ui.container.GridLayout
+        LeftPanel                matlab.ui.container.Panel
+        LoadCMPdataButton        matlab.ui.control.Button
+        HyperbolafittingLabel    matlab.ui.control.Label
+        SavevelocitiesButton     matlab.ui.control.Button
+        t0nsSlider               matlab.ui.control.Slider
+        t0nsSliderLabel          matlab.ui.control.Label
+        vcmnsSlider              matlab.ui.control.Slider
+        vcmnsSliderLabel         matlab.ui.control.Label
+        PickApex                 matlab.ui.control.Button
+        CorrectT0_Button         matlab.ui.control.Button
+        dxoffset_EditField       matlab.ui.control.EditField
+        dxoffsetmonesidedLabel   matlab.ui.control.Label
+        Startoffset_EditField    matlab.ui.control.EditField
+        StartoffsetTxRxmLabel    matlab.ui.control.Label
+        ProfileListBox           matlab.ui.control.ListBox
+        ProfileListBoxLabel      matlab.ui.control.Label
+        PlotButtonBox            matlab.ui.container.ButtonGroup
+        AspectRatioListBox       matlab.ui.control.ListBox
+        AspectRatioListBoxLabel  matlab.ui.control.Label
+        ScaleEditField           matlab.ui.control.EditField
+        ScaleEditFieldLabel      matlab.ui.control.Label
+        GrayscaleButton          matlab.ui.control.RadioButton
+        WiggleButton             matlab.ui.control.RadioButton
+        RightPanel               matlab.ui.container.Panel
+        HowTo_Label              matlab.ui.control.Label
+        SemblanceanalysisPanel   matlab.ui.container.Panel
+        calculateSemblanceButton matlab.ui.control.Button
+        SavevelocitiesButton_2   matlab.ui.control.Button
+        ManualPick_Button        matlab.ui.control.Button
+        ThresholdEditField       matlab.ui.control.EditField
+        ThresholdEditFieldLabel  matlab.ui.control.Label
+        epsEditField       matlab.ui.control.EditField
+        epsEditFieldLabel  matlab.ui.control.Label
+        minptsEditField       matlab.ui.control.EditField
+        minptsEditFieldLabel  matlab.ui.control.Label
+        AutoPick_Button          matlab.ui.control.Button
+        vMax_EditField           matlab.ui.control.EditField
+        v_maxcmnsEditFieldLabel  matlab.ui.control.Label
+        vMin_EditField           matlab.ui.control.EditField
+        v_mincmnsEditFieldLabel  matlab.ui.control.Label
+        UISemblance              matlab.ui.control.UIAxes
+        UIData                   matlab.ui.control.UIAxes
+    end
+
+    % Properties that correspond to apps with auto-reflow
+    properties (Access = private)
+        onePanelWidth = 576;
+        data
+    end
+
+    % Callbacks that handle component events
+    methods (Access = private)
+
+        % Button pushed function: LoadCMPdataButton
+        function LoadData_cb(app, event)
+
+            % get folder name with data
+            if ispc
+                if exist('temp.temp','file') % read last opened folder from temp.temp
+                    fid=fopen('temp.temp','r');
+                    fn=textscan(fid,'%s');
+                    fclose(fid);
+                    if ~isempty(fn{1})
+                        pfad=uigetdir(fn{1}{1},'Choose folder with radargrams');
+                    else
+                        pfad=uigetdir([],'Choose folder with radargrams');
[…]
