### 2024-09-19 eb3563e
corrected small bug in sgy-export for DZT-projects with different dt
