+++ GSSI_Import/DZT_Convert.m
@@ -35,7 +35,7 @@ removeOutliers=0; % do you want to remove coordinate outliers?
 
 % Export to other formats
 export2mat=1; % export to Multichannel-GPR format for radargrams (mat-files)
-export2segy=0; % export all radargrams as segy-files
+export2segy=1; % export all radargrams as segy-files
 constoff=0; % if=1: a constant coordinate offset will be subtracted and coordinates will be in mm accuracy in segy file (offsets will be saved in Inline3D (x) and Crossline3D (y))
 
 
@@ -841,6 +841,10 @@ if export2segy==1
     if isfield(h,'dt2') % is DF antenna
         h.dt=h.dt*1e9; % convert to ns
         h.dt2=h.dt2*1e9;
+        for i=1:length(headers)
+            headers{i}.dt=headers{i}.dt*1e9;
+            headers{i}.dt2=headers{i}.dt2*1e9;
+        end
     end
     
     for i=1:length(radargrams)
@@ -848,9 +852,9 @@ if export2segy==1
             global_coords{i}(:,3)=zeros(size(global_coords{i}(:,1))); % if no topography present, set to zero
         end
         if chan(i)==2 && isfield(h,'dt2')
-            export2sgy2D(radargrams{i},h.dt2,global_coords{i}(:,1),global_coords{i}(:,2),fullfile(pfad,'SEGY',[listrad.name{i},'_Chan',int2str(listrad.chan(i)),'.sgy']),global_coords{i}(:,3),constoff);
+            export2sgy2D(radargrams{i},headers{i}.dt2,global_coords{i}(:,1),global_coords{i}(:,2),fullfile(pfad,'SEGY',[listrad.name{i},'_Chan',int2str(listrad.chan(i)),'.sgy']),global_coords{i}(:,3),constoff);
         else
-            export2sgy2D(radargrams{i},h.dt,global_coords{i}(:,1),global_coords{i}(:,2),fullfile(pfad,'SEGY',[listrad.name{i},'_Chan',int2str(listrad.chan(i)),'.sgy']),global_coords{i}(:,3),constoff);
+            export2sgy2D(radargrams{i},headers{i}.dt,global_coords{i}(:,1),global_coords{i}(:,2),fullfile(pfad,'SEGY',[listrad.name{i},'_Chan',int2str(listrad.chan(i)),'.sgy']),global_coords{i}(:,3),constoff);
         end
     end
     
