### 2024-09-19 tinawunderlich/MultichannelGPR#1 @lrverdon
Problem finding .temp files
When following the Tutorial as part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/6767), in different steps of the tutorial, I had problems with the .temp files which the system could not find, e.g. when running the scripts 'DZT_convert.m' (dzt.temp), 'processing.m', 'Radargrams2bins.m', 'make_timeslices.m' (temp.temp), 'saveFigures.m' (radtemp.temp).
This occurred during different tests on different computers, but did not occur after I had removed a previous version of MultichannelGPR which had been in the Matlab Search Path.
If more information is available on what may cause this (older version of MultichannelGPR in the search path?), it may be useful to mention this in the documentation to remind the users.

I ran the tests on PCs with Windows 10 Pro 64 bits, Version 22H2
