### 2024-09-19 tinawunderlich/MultichannelGPR#2 @lrverdon
Problem when running isochrone migration in 'Process_SingleProfile' script
When running 'Process_SingleProfile.m' (step 5 of the Tutorial) as part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/6767), I could not get the isochrone migration to work on the basis of the description of the 'Process_SingleProfile' script in the tutorial and in the 'Scripts_for_2D' part  of the Wiki.
I selected the .mat files with the time and the velocity (should these be vectors with length equal to the number of temporal samples in the profile?) but the function kept running without end or gave an error message like the one below:

Arrays have incompatible sizes for this operation.
Error in isochrone_mig_2d_varV (line 57)
    v=zeros(size(data))+v;
Error in processingTestGUI_mat>processing (line 1132)
        [datatemp,zmig]=isochrone_mig_2d_varV(datatraces,x,tp,vgrid,params.aperture,0);
Error in processingTestGUI_mat/apply_call (line 967)
        [S.proc,S.ns,S.tproc,S.zmig,S.f,S.ampl]=processing(steps,order,S.raw,S.xprof,S.xyz,S.t,params,S);
Error while evaluating UIControl Callback.

Perhaps it could be further clarified in the documentation how to use the isochrone migration in 'Process_SingleProfile.m'? When I ran the 'isochrone_mig_2d_varV' function manually on some test GPR data, it worked fine.
(tested on a PC with Windows 10 Pro 64 bits, Version 22H2)
