### 2024-09-20 tinawunderlich/MultichannelGPR#3 @lrverdon
Question about profiles in 'profiles2mat' folder when dealing with Spidar data
During the review for JOSS (https://github.com/openjournals/joss-reviews/issues/6767) I carried out a test with a small set of Sensors & Software Spidar data (the raw data have names like 'NIC01_Line001.DT1' and 'NIC01_Line001.HD' until 'NIC07_Line052.DT1' and 'NIC07_Line052.HD'). So there are 52 swaths with 7 channels each.
When running 'Spidar_processing.m', there are as many raw .mat profiles (in the folder \profiles2mat) as there are swaths (52), and their number of traces is equal to the total number of traces in the swath (so the number of traces of one profile x 7). They have names like 'NIC_Line_1.mat' until 'NIC_Line_52.mat').
The same is true for the processed profiles (in the folder \profiles2mat\proc; with the same names as the raw .mat files).
Is this how it is meant to be?
In the 'radargrams.mat' file the number of profiles (364) and the number of traces per profile are correct. The creation of time slices using 'Bin_SpidarProc.m' and 'make_Timeslices' also happened correctly.

(Tested on a PC with Windows 10 Pro 64 bits, Version 22H2)
