### 2024-09-20 tinawunderlich/MultichannelGPR#3 @tinawunderlich
Yes, this is how it is meant to be. This is due to the fact that for large datasets it is not possible to save all profiles with all channels in one radargrams.mat-file. Therefore (and similar to the Mala MIRA data handling that I decided on during the beginning of this software development) there are these individual swaths in the profiles2mat and profiles2mat/proc folder. For smaller datasets as yours it is also possible to export them (=join the profiles) as one single radargrams.mat (as you also tried obviously).
