### 2024-09-20 tinawunderlich/MultichannelGPR#4 @lrverdon
Statement of need
One of the review criteria for JOSS (https://github.com/openjournals/joss-reviews/issues/6767) is a statement of need in the documentation.
In the Wiki (on the 'Home' page) it is explained what MultichannelGPR is and is not, but I could not find a more explicit description of which problems or shortcomings in the existing GPR software packages MultichannelGPR intends to
solve (perhaps I overlooked this). This could be stated in a few sentences.
A clear statement of need is already given in the JOSS paper, so the statement of need in the Wiki could be a summary of the one in the paper.
