### 2024-09-20 tinawunderlich/MultichannelGPR#1 @tinawunderlich
Thank you for testing this. Unfortunately, this is an issue not easy to solve. The .temp-files are saved for remembering the last given folder so that when you run a script multiple times you do not need to click through the complete folder structure.
Normally I am working on a Mac, but at some point I was forced to use a Windows PC for programming. At this specific PC it occured that there was a problem with creating/reading these files as there seems to be a special thing on how to create hidden files. Unfortunately I cannot use that PC anymore. At that point I added the "if ispc..." part in the beginning when reading/creating the temp files. Obviously, some Windows versions have a problem with this. I'll think about this problem and how to solve it in the next few days.
There is already a small hint in the Troubleshooting-part of the Wiki, but maybe I should post it more pronounced, and also in the Tutorial.
I'll come back to this issue...
