### 2024-09-20 tinawunderlich/MultichannelGPR#2 @tinawunderlich
Thank you for testing this! I have to admit that this is a function I seldom use. I'll have a look and fix it. Maybe I could limit the isochrone migration in the Process-SingleProfile to a constant velocity (as for the topomigration) as this would be much easier to use without the need to create the v.mat and vt.mat-files before. If a user wants to use a more complex velocity function this can be applied in the Processing.m. Do you think that this is sufficient for the Process_SingleProfile?
