### 2024-09-20 tinawunderlich/MultichannelGPR#4 @tinawunderlich
Thanks for this suggestion! I added a further paragraph on this from the paper into the Wiki.
