### 2024-09-20 tinawunderlich/MultichannelGPR#1 @tinawunderlich
For me to better understand the problem, could you reproduce the error and post the error message? Thanks!
