### 2024-09-23 tinawunderlich/MultichannelGPR#1 @lrverdon
The errors which I got when MultichannelGPR_v1.1.3 was still in my Matlab Search Path were of the form (example when running 'Processing.m', step 5 of Tutorial):

Error using fileattrib
The system cannot find the file specified
Error in Processing (line 43)
        fileattrib('radtemp.temp','-h');

For later tests I have removed MultichannelGPR_v1.1.3, and I have not entered MultichannGPR_v1.2.4 in the Matlab Search Path, I just opened the scripts in the folder where they are stored, and ran them as indicated in the Tutorial. I no longer get the error above, but when going through the tutorial recently, I had the following error in two different steps of the tutorial:

1) When running the script 'Process_SingleProfile.m' (step 5). I got the error:

Error using matlab.ui.internal.dialog.FileSystemChooser/PathParser (line 87)
Invalid directory to operate on
Error in matlab.ui.internal.dialog.FileSystemChooser/set.InitialPathName (line 32)
                iPath = PathParser(obj,iPath);
Error in uigetdir_helper (line 36)
    dirdlg.InitialPathName = dialog_pathname;
Error in uigetdir (line 68)
[directoryname] = uigetdir_helper(varargin{:});
Error in Process_SingleProfile (line 27)
            folder=uigetdir(fn{1}{1},'Select data-folder');

I could avoid this by commenting out line 21-35 of the script, then the error below appeared:

Error using fprintf
Invalid file identifier. Use fopen to generate a valid file identifier.
Error in Process_SingleProfile (line 38)
        fprintf(fid,'%s',folder);

Also commenting out line 37-40 allowed selecting the 'setCoords' folder.

2) I got the same error when running 'saveFigures.m' (step 12)

Error using matlab.ui.internal.dialog.FileSystemChooser/PathParser (line 87)
Invalid directory to operate on
Error in matlab.ui.internal.dialog.FileSystemChooser/set.InitialPathName (line 32)
                iPath = PathParser(obj,iPath);
Error in uigetdir_helper (line 36)
    dirdlg.InitialPathName = dialog_pathname;
Error in uigetdir (line 68)
[directoryname] = uigetdir_helper(varargin{:});
Error in saveFigures (line 60)
            pfad_rad=uigetdir(fn{1}{1},'Choose folder with radargrams');

These two errors refer to lines in the scripts dealing with the .temp files, so I assume they are a related problem.

Perhaps one part of the solution to avoid errors in case anything goes wrong may be to build in an 'escape route'
when the system cannot find the temporary file or the value in the temporary file is wrong, for example in that case the current Matlab folder is given as default in the folder selection dialog.
