### 2024-09-23 tinawunderlich/MultichannelGPR#2 @lrverdon
Yes, I think that would be sufficient. Then the user gets an impression of what the data looks like after migration (even if there are variations in the velocity, a good average will be a nice approximation). In case certain mat-files can be selected, it is useful to specify in the guidelines if they can/should be scalars, vectors or matrices with a certain size.
