### 2024-09-23 tinawunderlich/MultichannelGPR#1 @tinawunderlich
Thank you very much for reproducing the error messages! I agree that the errors are relatde to the temp-files. I'll try to build an "escape-route" as you suggested, hoping that I can think of any possible problem that could occur ;-)
