### 2024-09-23 tinawunderlich/MultichannelGPR#2 @tinawunderlich
Ok, thanks. Then I'll implement it like that.
