### 2024-09-24 65f4c83
in Process_SingleProfile: for migration only a constant velocity can be used (easier input) in this test tool. And corrected some bugs when changing between raw/proc data.
