+++ GUIs/processingTestGUI_mat.m
@@ -1,8 +1,9 @@
 function []=processingTestGUI_mat(radargrams,t,x,global_coords,folder)
 
 % GUI for plotting of profiles and interactive testing of processing steps
+% for mat-files
 %
-% Dr. Tina Wunderlich, CAU Kiel 2023, tina.wunderlich@ifg.uni-kiel.de
+% Dr. Tina Wunderlich, CAU Kiel 2023-2024, tina.wunderlich@ifg.uni-kiel.de
 %
 % Input:
 % radargrams, t, x, global_coords, folder
@@ -209,23 +210,21 @@ S.fh.Visible='on';
         
         % 5th column:
         off5=110;
-        % isochrone migration 2d (reduced to 1d-v-function for easier input)
+        % isochrone migration 2d (reduced to constant v for easier input)
         S.migration=uicontrol(S.fh,'Style','checkbox','String','Isochrone migration v(z)','FontWeight','bold','Position',[1100-off1-off5 300 200 15],'Value',0,'Callback',@migration_call);
-        S.v=uicontrol(S.fh,'Style','edit','String','','Enable','off','Position',[1120-off1-off5 270 35 20],'Callback',@readv_call);
-        S.tv=uicontrol(S.fh,'Style','edit','String','','Enable','off','Position',[1120-off1-off5 240 35 20],'Callback',@readtv_call);
-        S.aperture=uicontrol(S.fh,'Style','edit','String','30','Enable','off','Position',[1120-off1-off5 210 35 20]);
+        S.v=uicontrol(S.fh,'Style','edit','String','0.1','Enable','off','Position',[1120-off1-off5 270 35 20]);
+        S.aperture=uicontrol(S.fh,'Style','edit','String','30','Enable','off','Position',[1120-off1-off5 240 35 20]);
         S.vtext=uicontrol(S.fh,'Style','text','String','v [m/ns]','Position',[1160-off1-off5 270 140 20],'HorizontalAlignment','left');
-        S.tvtext=uicontrol(S.fh,'Style','text','String','corresp. t [ns]','Position',[1160-off1-off5 240 140 20],'HorizontalAlignment','left');
-        S.aperturetext=uicontrol(S.fh,'Style','text','String','aperture [°]','Position',[1160-off1-off5 210 140 20],'HorizontalAlignment','left');
+        S.aperturetext=uicontrol(S.fh,'Style','text','String','aperture [°]','Position',[1160-off1-off5 240 140 20],'HorizontalAlignment','left');
         
         % Topomigration/korrektur (reduced to constant v for easier input)
-        S.topo=uicontrol(S.fh,'Style','checkbox','String','Topographic migration (const. v)','FontWeight','bold','Position',[1100-off1-off5 180 200 15],'Value',0,'Callback',@topo_call);
-        S.v2=uicontrol(S.fh,'Style','edit','String','0.1','Enable','off','Position',[1120-off1-off5 150 35 20]);
-        S.flagtopo=uicontrol(S.fh,'Style','edit','String','1','Enable','off','Position',[1120-off1-off5 120 35 20]);
-        S.aperture2=uicontrol(S.fh,'Style','edit','String','30','Enable','off','Position',[1120-off1-off5 90 35 20]);
-        S.v2text=uicontrol(S.fh,'Style','text','String','v [m/ns]','Position',[1160-off1-off5 150 140 20],'HorizontalAlignment','left');
-        S.flagtopotext=uicontrol(S.fh,'Style','text','String','correction(1), migration(2)','Position',[1160-off1-off5 120 140 20],'HorizontalAlignment','left');
-        S.aperture2text=uicontrol(S.fh,'Style','text','String','aperture [°]','Position',[1160-off1-off5 90 140 20],'HorizontalAlignment','left');
[…]
