### 2024-09-24 0a3ee4a
adjusted Mala3D and MalaProcessing so that they are compatible with settings from Process_SingleProfile and allowed different input formats for velocity (constant as a number or a file name)
