+++ Mala_Import/MalaProcessing.m
@@ -22,8 +22,8 @@ clc
 platform=2; % Linux=1, Mac=2, Windows=3
 
 % Select number of profiles:
-profile_min=0;  % minimum profile number
-profile_max=95;  % maximum profile number
+profile_min=1;  % minimum profile number
+profile_max=2;  % maximum profile number
 % number of channels for this dataset
 channels=16; % number of channels
 
@@ -39,7 +39,7 @@ settings='settings.txt';
 
 %%% Raw data is only read if userawdata=0! Otherwise mat-files of raw
 %%% data in folder profiles2mat are used.
-userawdata=1;  % if =1: use aready read in raw data and apply new processing steps
+userawdata=0;  % if =1: use aready read in raw data and apply new processing steps
 
 % save radargrams.mat of processed profile data in one variable (requires large memory -> probably not
 % working for every data set)?
@@ -122,14 +122,16 @@ if ~exist(fullfile(foldername,settings),'file') % if no settings-file is found: 
 
     fid=fopen(fullfile(foldername,settings),'wt');
     fprintf(fid,'do_amplitudeOffset 1\ntstart[ns] 0\ntend[ns] 100\n\n');
-    fprintf(fid,'do_traceInterpolation 0\nminfactor 3\nmaxfactor 3\n\n');
+    fprintf(fid,'do_badTraceRemoval 0\nminfactor 3\nmaxfactor 3\n\n');
+    fprintf(fid,'do_constTraceDist 2\ndx[m] 0.02\n\n');
+    fprintf(fid,'do_traceInterpolation 3\ngap 10\n\n');
+    fprintf(fid,'do_medfilt 0\nnumsamp 3\n\n');
+    fprintf(fid,'do_medfilt_x 0\nnumsamp_x 3\ntstart_x 0\n\n');
     fprintf(fid,'do_sphericalDivergence 0\n\n');
     fprintf(fid,'do_attenuationCorrection 0\nsigma[S/m] 0.002\neps 15\n\n');
     fprintf(fid,'do_spectralWhitening 0\nfmin_sw[MHz] 100\nfmax_sw[MHz] 600\nalpha 0.01\n\n');
     fprintf(fid,'do_bandpass 5\nfstart[MHz] 100\nfend[MHz] 600\n\n');
-    fprintf(fid,'do_constTraceDist 2\ndist[m] 0.02\ngap 100\n\n');
-    fprintf(fid,'do_medianFilter 0\nm 3\n\n');
-    fprintf(fid,'do_migration2d_vz 0\nv-file[m/ns] \ntv-file[ns] \naperture_m[degree] 30\n\n');
+    fprintf(fid,'do_migration2d_vz 0\nv-file[m/ns] vgrid.mat\ntv-file[ns] vt.mat\naperture_m[degree] 30\n\n');
     fprintf(fid,'do_topomig2d 0\nv[m/ns] 0.1\nflag 1\naperture_t[degree] 30\n\n');
     fprintf(fid,'do_cutTWT 6\ntmax[ns] 80\n\n');
     fprintf(fid,'do_khighpass 0\nkcutoff[1/m] 0.1\n\n');
@@ -150,7 +152,11 @@ else
     temp=textscan(fid,'%s%s');
     fclose(fid);
     % get order of processing steps:
[…]
