+++ Mala_Import/Mala3D.m
@@ -19,10 +19,10 @@ clc
 platform=2; % Linux=1, Mac=2, Windows=3
 
 % Select number of profiles:
-profile_min=3;  % minimum profile number
-profile_max=6;  % maximum profile number
+profile_min=1;  % minimum profile number
+profile_max=2;  % maximum profile number
 % number of channels for this dataset
-channels=32; % number of channels
+channels=16; % number of channels
 
 changeDir=0; % if =1: change the sign of the y-antenna-GPS-offset, if =0: use offsets as written in file
 
@@ -170,17 +170,19 @@ if ~exist(fullfile(foldername,settings),'file') % if no settings-file is found: 
     fprintf(fid,'do_constTraceDist 2\ndx[m] 0.02\n\n');
     fprintf(fid,'do_traceInterpolation 3\ngap 10\n\n');
     fprintf(fid,'do_medfilt 0\nnumsamp 3\n\n');
+    fprintf(fid,'do_medfilt_x 0\nnumsamp_x 3\ntstart_x 0\n\n');
     fprintf(fid,'do_sphericalDivergence 0\n\n');
     fprintf(fid,'do_attenuationCorrection 0\nsigma[S/m] 0.002\neps 15\n\n');
     fprintf(fid,'do_spectralWhitening 0\nfmin_sw[MHz] 100\nfmax_sw[MHz] 600\nalpha 0.01\n\n');
     fprintf(fid,'do_bandpass 6\nfstart[MHz] 100\nfend[MHz] 600\n\n');
-    fprintf(fid,'do_migration2d_vz 0\nv-file[m/ns] \ntv-file[ns] \naperture_m[degree] 30\n\n');
+    fprintf(fid,'do_migration2d_vz 0\nv-file[m/ns] vgrid.mat\ntv-file[ns] vt.mat\naperture_m[degree] 30\n\n');
     fprintf(fid,'do_topomig2d 0\nv[m/ns] 0.1\nflag 1\naperture_t[degree] 30\n\n');
     fprintf(fid,'do_cutTWT 0\ntmax[ns] 80\n\n');
     fprintf(fid,'do_khighpass 0\nkcutoff[1/m] 0.1\n\n');
     fprintf(fid,'do_applygain 0\ng1 -20\ng2 0\ng3 10\ng4 20\ng5 30\n\n');
     fprintf(fid,'do_normalization 0\nqclip 0.98\n\n');
     fprintf(fid,'do_removeMeanMedianTrace 0\nmeanmedian 1\nnumtraces 0\n\n');
+    fprintf(fid,'do_t0shift 0\nt0s[ns] 5.0\n\n');
     fprintf(fid,'do_t0CorrectionThreshold 0\nthreshold -1\n\n');
     fprintf(fid,'do_t0CorrectionReferencetrace 0\nprofilenumber 0\nchannelnumber 1\ntracenumber 1\nt0[ns] 0.1\n\n');
     fprintf(fid,'do_crossprofileshift 5\nmaxshift_cps 30\n\n');
@@ -194,7 +196,11 @@ else
     temp=textscan(fid,'%s%s');
     fclose(fid);
     % get order of processing steps:
[…]
