### 2024-09-24 tinawunderlich/MultichannelGPR#2 @tinawunderlich
It is implemented in the GUI and Mala3D & MalaProcessing are adjusted to the settings-file from Process_SingleProfile. When you are using mat-files the required format is given in the Wiki.
