+++ SensorsSoftware_Import/Bin_SpidarProc.m
@@ -40,25 +40,26 @@ warning('off');
 if ispc
     if exist('temp.temp') % read last opened folder from temp.temp
         fid=fopen('temp.temp','r');
-        fn=textscan(fid,'%s');
+        if fid~=-1
+            fn=textscan(fid,'%s');
+        else
+            fn{1}=[];
+        end
         fclose(fid);
         if ~isempty(fn{1})
             foldername=uigetdir(fn{1}{1},'Choose folder with original data files');
         else
             foldername=uigetdir([],'Choose folder with original data files');
         end
-        fileattrib('temp.temp','-h');
         fid=fopen('temp.temp','wt');
         fprintf(fid,'%s',foldername);
         fclose(fid);
-        fileattrib('temp.temp','+h');
     else
         foldername=uigetdir([],'Choose folder with original data files'); % path to radargram-folder
 
         fid=fopen('temp.temp','wt');
         fprintf(fid,'%s',foldername);
         fclose(fid);
-        fileattrib('temp.temp','+h');
     end
 else
     if exist('.temp.temp') % read last opened folder from temp.temp
