+++ Visualization/Tsl_Radargrams.m
@@ -31,25 +31,27 @@ if ~ispc; menu('Choose folder with timeslices','OK'); end
 if ispc
     if exist('tsltemp.temp') % read last opened folder from temp.temp
         fid=fopen('tsltemp.temp','r');
-        fn=textscan(fid,'%s');
+        if fid~=-1
+            fn=textscan(fid,'%s');
+        else
+            fn{1}=[];
+        end
         fclose(fid);
         if ~isempty(fn{1})
             pfad_tsl=uigetdir(fn{1}{1},'Choose folder with timeslices');
         else
             pfad_tsl=uigetdir([],'Choose folder with timeslices');
         end
-        fileattrib('tsltemp.temp','-h');
+
         fid=fopen('tsltemp.temp','wt');
         fprintf(fid,'%s',pfad_tsl);
         fclose(fid);
-        fileattrib('tsltemp.temp','+h');
     else
         pfad_tsl=uigetdir([],'Choose folder with timeslices'); % path to radargram-folder
 
         fid=fopen('tsltemp.temp','wt');
         fprintf(fid,'%s',pfad_tsl);
         fclose(fid);
-        fileattrib('tsltemp.temp','+h');
     end
 else
     if exist('.tsltemp.temp') % read last opened folder from temp.temp
@@ -76,25 +78,26 @@ if ~ispc; menu('Choose folder with radargrams','OK'); end
 if ispc
     if exist('radtemp.temp') % read last opened folder from temp.temp
         fid=fopen('radtemp.temp','r');
-        fn=textscan(fid,'%s');
+        if fid~=-1
+            fn=textscan(fid,'%s');
+        else
+            fn{1}=[];
+        end
         fclose(fid);
         if ~isempty(fn{1})
             pfad_rad=uigetdir(fn{1}{1},'Choose folder with radargrams');
         else
             pfad_rad=uigetdir([],'Choose folder with radargrams');
         end
-        fileattrib('radtemp.temp','-h');
         fid=fopen('radtemp.temp','wt');
         fprintf(fid,'%s',pfad_rad);
         fclose(fid);
-        fileattrib('radtemp.temp','+h');
     else
         pfad_rad=uigetdir([],'Choose folder with radargrams'); % path to radargram-folder
 
         fid=fopen('radtemp.temp','wt');
         fprintf(fid,'%s',pfad_rad);
         fclose(fid);
-        fileattrib('radtemp.temp','+h');
     end
 else
     if exist('.radtemp.temp') % read last opened folder from temp.temp
