+++ Mala_Import/Mala3D.m
@@ -76,25 +76,26 @@ end
 if ispc
     if exist('temp.temp') % read last opened folder from temp.temp
         fid=fopen('temp.temp','r');
-        fn=textscan(fid,'%s');
+        if fid~=-1
+            fn=textscan(fid,'%s');
+        else
+            fn{1}=[];
+        end
         fclose(fid);
         if ~isempty(fn{1})
             foldername=uigetdir(fn{1}{1},'Choose rSlicer folder');
         else
             foldername=uigetdir([],'Choose rSlicer folder');
         end
-        fileattrib('temp.temp','-h');
         fid=fopen('temp.temp','wt');
         fprintf(fid,'%s',foldername);
         fclose(fid);
-        fileattrib('temp.temp','+h');
     else
         foldername=uigetdir([],'Choose rSlicer folder'); % path to radargram-folder
 
         fid=fopen('temp.temp','wt');
         fprintf(fid,'%s',foldername);
         fclose(fid);
-        fileattrib('temp.temp','+h');
     end
 else
     if exist('.temp.temp') % read last opened folder from temp.temp

[…]
