### 2024-10-14 tinawunderlich/MultichannelGPR#1 @tinawunderlich
I changed now all scripts and for Windows the temporarily saved files are now not hidden (which was the main source of problems). If you still encounter problems with uigetfile/uigetdir there is a solution in the Wiki/Troubleshooting, which is also linked to the tutorial.
