### 2024-10-14 76e0643
updating of the manuscript for JOSS, including an updated figure and bibliography
