+++ paper/paper.md
@@ -13,7 +13,7 @@ authors:
 affiliations:
  - name: Institute of Geosciences, Kiel University, Kiel, Germany
    index: 1
-date: 17 January 2024
+date: 14 October 2024
 bibliography: paper.bib
 
 
@@ -21,18 +21,18 @@ bibliography: paper.bib
 
 # Summary
 
-Ground Penetrating Radar (GPR) is a popular geophysical method for near surface investigations using high-frequency electromagnetic waves for a broad range of applications such as archaeological prospection [e.g. @Trinks:2018], hydrology [e.g. @Annan:2005], urban infrastructure detection [e.g. @Pajewski:2019], geological mapping [e.g. @Buck:2024], permafrost investigations [e.g. @Westermann:2010] or soil science [e.g. @Gates:2023]. Pulsed electromagnetic waves are transmitted from a transmitter antenna, penetrate into the ground and are reflected and transmitted at interfaces until the reflected waves are recorded by a receiving antenna. For mapping, the transmitter and receiver antenna (=one channel) are mostly carried with a constant small offset between them along a profile. For applications requiring high spatial resolution multi-channel equipment is available collecting several parallel profiles with small profile spacing (e.g. 4 cm or 8 cm) in one swath. 
+Ground Penetrating Radar (GPR) is a popular geophysical method for near surface investigations using high-frequency electromagnetic waves for a broad range of applications such as archaeological prospection [e.g. @Trinks:2018], hydrology [e.g. @Annan:2005], urban infrastructure detection [e.g. @Pajewski:2019], geological mapping [e.g. @Buck:2024], permafrost investigations [e.g. @Westermann:2010] or soil science [e.g. @Gates:2023]. Pulsed electromagnetic waves are transmitted from a transmitter antenna, penetrate into the ground and are reflected and transmitted at interfaces until the reflected waves are recorded by a receiving antenna. For mapping, the transmitter and receiver antenna (=one channel) are mostly carried with a constant small offset between them along a profile. For applications requiring high spatial resolution multi-channel equipment is available collecting several parallel profiles with small profile spacing (e.g. 4 cm or 8 cm, depending on the frequency) in one swath. 
 
-`MultichannelGPR` is a structured collection of MATLAB scripts and functions for processing, visualization and export of single- and multi-channel GPR data. Although it was originally intended and developed for the processing of multichannel Malå MIRA data, it evolved to a more general processing tool also for data from other systems, e.g. GSSI, and partially Sensors&Software and Radarteam. But as the needed input format is clearly defined, the user can input any GPR data from any system, if the data is converted into the required mat-format.
+`MultichannelGPR` is a structured collection of MATLAB scripts and functions for processing, visualization and export of single- and multi-channel GPR data. Although it was originally intended and developed for the processing of multichannel Malå MIRA data, it evolved to a more general processing tool also for data from other systems, e.g. GSSI, and partially Sensors&Software and Radarteam. But as the needed input format is clearly defined, the user can input any GPR data from any system, if the data is converted into the required mat-format before usage of `MultichannelGPR`.
 It is structured into various folders containing MATLAB scripts for specific purposes that are mainly data import, processing in 2D and 3D (or cross-channel for multichannel data), visualization and velocity analysis. Some Graphical User Interfaces (GUIs) for processing and visualization are also provided as well as export functionalities for georeferenced radargrams and timeslices for easy import into a Geographical Information System (GIS).
[…]
