+++ paper/paper.bib
@@ -201,3 +201,17 @@ DOI = {10.5194/tc-4-475-2010}
   journal = {GitHub repository},
   url = {https://github.com/tinawunderlich/MultichannelGPR}
 }
+
+@misc{Goodman,
+  author = {Goodman, D.},
+  title = {GPR-SLICE},
+  journal = {Website},
+  url = {https://www.gpr-survey.com/index-2.html}
+}
+
+@misc{Sandmeier,
+  author = {Sandmeier, K.-J.},
+  title = {REFLEXW},
+  journal = {Website},
+  url = {https://www.sandmeier-geo.de/reflexw.html}
+}
