### 2024-10-16 #6767 @LyceanEM
I have now worked through the tutorial examples for this package, and I am impressed with the variety of functionality on offer. In terms of the checklist @tinawunderlich there is not currently any documentation on contribution guidelines, and I have raised this as an issue on the repo. Aside from this I have no comments that haven't been mentioned by @lrverdon and addressed.
