### 2024-10-18 d46efca
Create CONTRIBUTING.md
