+++ CONTRIBUTING.md
@@ -1,0 +1,14 @@
+# CONTRIBUTING
+
+If you are interested in contributing to MultichannelGPR, there are several possibilities:
+
+## Bug reports
+Bug reports are very helpful in improving the program and are highly appreciated.
+When making bug reports in issues, please include:
+- a description of your problem: What did you expect to happen and what happened actually?
+- a description of the processing steps that you did before.
+- a description of your data (which equipment was used, was it measured with GNSS/total station/hand marker/...?)
+- which operating system you use.
+
+## Adding new features
+If you have an idea for a new feature, please raise an issue and describe your idea. We can then discuss if/how this can be included and who is going to do this.
