+++ GSSI_Import/DZT_Convert.m
@@ -11,7 +11,7 @@ clear all
 close all
 clc
 
-app='SIR4000';    % Equipment: SIR20 / SIR30 / SIR3000 / SIR4000 / Tablet
+app='SIR20';    % Equipment: SIR20 / SIR30 / SIR3000 / SIR4000 / Tablet
 
 dataplot=1; % plot radargram for controlling? 1=yes, 0=no
 
@@ -23,7 +23,7 @@ offsetGPS_Y=0; % [m] Offset between GPS and antenna midpoint in profile directio
 h_GPS=0; % [m] height of GPS/prism above ground
 
 % Options for calculating inline coordinates for each trace:
-coords_opt=2;   % =1: trace coordinate is difference to beginning of profile (only use this for straight profiles!)
+coords_opt=1;   % =1: trace coordinate is difference to beginning of profile (only use this for straight profiles!)
                 % =2: trace coordinates are calculated by taking the cumulative sum of the coordinate differences between subsequent traces (better for curvy profiles, but not useful for strong GPS-antenna movements)
 
 % Attention: still experimental! If you think that the output radargrams
@@ -35,7 +35,7 @@ removeOutliers=0; % do you want to remove coordinate outliers?
 
 % Export to other formats
 export2mat=1; % export to Multichannel-GPR format for radargrams (mat-files)
-export2segy=1; % export all radargrams as segy-files
+export2segy=0; % export all radargrams as segy-files
 constoff=0; % if=1: a constant coordinate offset will be subtracted and coordinates will be in mm accuracy in segy file (offsets will be saved in Inline3D (x) and Crossline3D (y))
 
 
