### 2024-10-18 #6767 @tinawunderlich
@LyceanEM thank you for your comments! I just added a CONTRIBUTING.md file to the repository and closed the issue.
