### 2024-10-21 #6767 @martinfleis
@tinawunderlich Please refer to the checklist [above](https://github.com/openjournals/joss-reviews/issues/6767#issuecomment-2425894169) for the pre-publication tasks.

On top of that, could you slightly edit the paper to avoid citing Wunderlich 2024, which is the repository that is being reviewed? It is a bit strange to cite the work itself in a paper about it.

Thank you!
