+++ paper/paper.bib
@@ -193,15 +193,6 @@ DOI = {10.5194/tc-4-475-2010}
 }
 
 
-@misc{fidgit,
-  author = {Wunderlich, T.},
-  title = {MultichannelGPR: A MATLAB tool for GPR data processing},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/tinawunderlich/MultichannelGPR}
-}
-
 @misc{Goodman,
   author = {Goodman, D.},
   title = {GPR-SLICE},
