### 2024-10-22 #6767 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
