### 2024-10-23 3f293f1
corrected references in paper.bib
