+++ paper/paper.bib
@@ -1,10 +1,10 @@
 @Inbook{Annan:2005,
-  author={Annan, A.P.},
-  editor={Rubin, Y. and Hubbard, S.S.},
-  title={GPR Methods for Hydrogeological Studies},
+  author={{Annan}, A.P.},
+  editor={{Rubin}, Y. and {Hubbard}, S.S.},
+  title={{GPR} {Methods} for {Hydrogeological} {Studies}},
   bookTitle={Hydrogeophysics},
   year={2005},
-  publisher={Springer Netherlands},
+  publisher={{Springer} {Netherlands}},
   address={Dordrecht},
   pages={185--213},
   isbn={978-1-4020-3102-1},
@@ -14,8 +14,8 @@
 
 @article{Buck:2024,
 author = {{Buck}, L. and {Bristow}, C.S.},
-title = {Using ground-penetrating radar to investigate deposits from the Storegga slide tsunami and other sand sheets in the Shetland Islands, UK},
-journal = {Journal of the Geological Society},
+title = {Using {Ground}-{Penetrating} {Radar} to investigate {Deposits} from the {Storegga} {Slide} {Tsunami} and other {Sand} {Sheets} in the {Shetland} {Islands}, {UK}},
+journal = {Journal of the {Geological} {Society}},
 volume = {181},
 number = {1},
 pages = {jgs2023-042},
@@ -24,9 +24,9 @@ doi = {10.1144/jgs2023-042}
 }
 
 @article{Corradini:2022,
-  title={Finding Mesolithic sites: a multichannel ground-penetrating radar (GPR) investigation at the ancient Lake Duvensee},
+  title={Finding {Mesolithic} {Sites}: {A} {Multichannel} {Ground}-{Penetrating} {Radar} ({GPR}) {Investigation} at the {Ancient} {Lake} {Duvensee}},
   author={{Corradini}, E. and {Gross}, D. and {Wunderlich}, T. and {L{\"u}bke}, H. and {Wilken}, D. and {Erkul}, E. and {Schm{\"o}lcke}, U. and {Rabbel}, W.},
-  journal={Remote Sensing},
+  journal={Remote {Sensing}},
   volume={14},
   number={3},
   pages={781},
@@ -36,9 +36,9 @@ doi = {10.1144/jgs2023-042}
 }
 
 @article{Corradini:2023,
-  title={A Day at the Bog: Preliminary Interpretation of Prehistoric Human Occupation at Ancient Lake Duvensee (Germany) by GPR Structures},
+  title={A {Day} at the {Bog}: {Preliminary} {Interpretation} of {Prehistoric} {Human} {Occupation} at {Ancient} {Lake} {Duvensee} ({Germany}) by {GPR} {Structures}},
   author={{Corradini}, E. and {Dreibrodt}, S. and {L{\"u}bke}, H. and {Schm{\"o}lcke}, U. and {Wieckowska-L{\"u}th}, M. and {Wunderlich}, T. and {Wilken}, D. and {Brozio}, J.P. and {Rabbel}, W.},
-  journal={Remote Sensing},
+  journal={Remote {Sensing}},
   volume={15},
   number={14},
   pages={3647},
@@ -49,8 +49,8 @@ doi = {10.1144/jgs2023-042}
 
 @article{Gates:2023,
 author = {{Gates}, Z.W. and {Galagedara}, L.W. and {Ziegler}, S.E.},
-title = {Combining ground penetrating radar methodologies enables large-scale mapping of soil horizon thickness and bulk density in boreal forests},
-journal = {Soil Use and Management},
[…]
