### 2024-10-24 #6767 @kthyng
@tinawunderlich If you'd like to join JOSS as a reviewer, please sign up here: https://reviewers.joss.theoj.org/!
